Doc website enhancements
- Feature emojis
- Site title
- Global flag page grammar fixes
- Linked NodeJS to getting started page
