+++ cli/package.json
@@ -2,7 +2,7 @@
   "name": "@serverless-ninja/sana",
   "package-name": "@serverless-ninja/sana",
   "description": "A command-line interface for analyzing an AWS account's serverless resources",
-  "version": "1.1.136",
+  "version": "1.1.137",
   "main": "app.mjs",
   "repository": "https://github.com/allanchua101/sana",
   "homepage": "https://allanchua101.github.io/sana/",
