+++ doc-website/index.md
@@ -1,11 +1,13 @@
 ---
 # https://vitepress.dev/reference/default-theme-home-page
 layout: home
+title: "sana"
+desc: "Analyze your AWS account's serverless resources"
 
 hero:
   name: "sana"
   text: "Analyze your AWS account's serverless resources"
-  tagline: Designed by AWS serverless engineers, for AWS serverless engineers
+  tagline: "Made by frugal engineers, for AWS serverless engineers."
   actions:
     - theme: brand
       text: Get Started
@@ -14,10 +16,13 @@ hero:
       text: View on Github
       link: https://github.com/allanchua101/sana
 features:
-  - title: Account-wide Scanning
+  - icon: 🌐
+    title: Account-wide Scanning
     details: Rapidly scan an entire account's serverless resources without leaving your VSCode
-  - title: Distribution Analysis
+  - icon: 📊
+    title: Distribution Analysis
     details: Uncover misconfigurations, outliers and anomalies using distribution analysis.
-  - title: Modern Web-based Documentation
+  - icon: ⚙️
+    title: Modern Web-based Documentation
     details: Easy to read web-based documentation built on top of modern tooling.
 ---
