+++ docs/assets/chunks/@localSearchIndexroot.fNJnBJgz.js
@@ -1,0 +1,1 @@
+const a='{"documentCount":91,"nextId":91,"documentIds":{"0":"/sana/dynamodb.html#sana-ddb","1":"/sana/dynamodb.html#example-usage","2":"/sana/dynamodb.html#kpi-specific-commands","3":"/sana/dynamodb/aggregation/ddb-count.html#sana-ddb-count","4":"/sana/dynamodb/aggregation/ddb-count.html#use-cases","5":"/sana/dynamodb/distribution/ddb-delete-protection-distribution.html#sana-ddb-delete-protection-distribution","6":"/sana/dynamodb/distribution/ddb-delete-protection-distribution.html#use-cases","7":"/sana/dynamodb/distribution/ddb-region-distribution.html#sana-ddb-region-distribution","8":"/sana/dynamodb/distribution/ddb-region-distribution.html#use-cases","9":"/sana/dynamodb/distribution/ddb-table-status-distribution.html#ddb-table-status-distribution","10":"/sana/full-service-analysis.html#full-service-analysis","11":"/sana/full-service-analysis.html#lambda-functions","12":"/sana/full-service-analysis.html#dynamodb-tables","13":"/sana/full-service-analysis.html#analyzing-multiple-services","14":"/sana/get-started.html#get-started","15":"/sana/get-started.html#running-your-first-scan","16":"/sana/get-started.html#next-steps","17":"/sana/global-flags.html#global-flags","18":"/sana/global-flags.html#profile-p","19":"/sana/global-flags.html#region-r","20":"/sana/global-flags.html#output-o","21":"/sana/global-flags.html#no-progress-bar","22":"/sana/global-flags.html#silent-mode","23":"/sana/installation.html#installation","24":"/sana/installation.html#pre-requisites","25":"/sana/installation.html#initial-installation","26":"/sana/installation.html#upgrading-an-existing-installation","27":"/sana/lambda.html#sana-lambda","28":"/sana/lambda.html#example-usage","29":"/sana/lambda.html#kpi-specific-commands","30":"/sana/lambda/aggregation/lambda-avg-ephemeral-storage-size.html#sana-lambda-avg-ephemeral-storage-size","31":"/sana/lambda/aggregation/lambda-avg-ephemeral-storage-size.html#more-examples","32":"/sana/lambda/aggregation/lambda-avg-memory-size.html#sana-lambda-avg-memory","33":"/sana/lambda/aggregation/lambda-avg-package-size.html#sana-lambda-avg-package-size","34":"/sana/lambda/aggregation/lambda-avg-package-size.html#use-cases","35":"/sana/lambda/aggregation/lambda-avg-timeout.html#sana-lambda-avg-timeout","36":"/sana/lambda/aggregation/lambda-count.html#sana-lambda-count","37":"/sana/lambda/aggregation/lambda-count.html#use-cases","38":"/sana/lambda/aggregation/lambda-max-package-size.html#sana-lambda-max-package-size","39":"/sana/lambda/aggregation/lambda-max-package-size.html#use-cases","40":"/sana/lambda/aliases/logging-kpis.html#sana-lambda-logging-kpis","41":"/sana/lambda/aliases/logging-kpis.html#combined-commands","42":"/sana/lambda/aliases/vpc-kpis.html#sana-lambda-vpc-kpis","43":"/sana/lambda/aliases/vpc-kpis.html#combined-commands","44":"/sana/lambda/distribution/lambda-app-log-lvl-distribution.html#sana-lambda-app-log-lvl-distribution","45":"/sana/lambda/distribution/lambda-app-log-lvl-distribution.html#configuration-purpose","46":"/sana/lambda/distribution/lambda-app-log-lvl-distribution.html#use-cases","47":"/sana/lambda/distribution/lambda-app-log-lvl-distribution.html#analyze-logging-kpis","48":"/sana/lambda/distribution/lambda-app-log-lvl-distribution.html#tips","49":"/sana/lambda/distribution/lambda-architecture-distribution.html#sana-lambda-architecture-distribution","50":"/sana/lambda/distribution/lambda-architecture-distribution.html#use-cases","51":"/sana/lambda/distribution/lambda-architecture-distribution.html#tips","52":"/sana/lambda/distribution/lambda-architecture-distribution.html#related-benchmarks-and-materials","53":"/sana/lambda/distribution/lambda-attached-layer-count-distribution.html#sana-lambda-attached-layer-count-distribution","54":"/sana/lambda/distribution/lambda-dlq-distribution.html#sana-lambda-dlq-distribution","55":"/sana/lambda/distribution/lambda-dlq-distribution.html#use-cases","56":"/sana/lambda/distribution/lambda-dlq-distribution.html#notes","57":"/sana/lambda/distribution/lambda-ephemeral-storage-distribution.html#sana-lambda-ephemeral-storage-distribution","58":"/sana/lambda/distribution/lambda-ephemeral-storage-distribution.html#use-cases","59":"/sana/lambda/distribution/lambda-log-format-distribution.html#sana-lambda-log-format-distribution","60":"/sana/lambda/distribution/lambda-log-format-distribution.html#configuration-purpose","61":"/sana/lambda/distribution/lambda-log-format-distribution.html#use-cases","62":"/sana/lambda/distribution/lambda-log-format-distribution.html#analyze-logging-kpis","63":"/sana/lambda/distribution/lambda-log-format-distribution.html#tips","64":"/sana/lambda/distribution/lambda-log-format-distribution.html#related-articles","65":"/sana/lambda/distribution/lambda-memory-distribution.html#sana-lambda-memory-distribution","66":"/sana/lambda/distribution/lambda-memory-distribution.html#use-cases","67":"/sana/lambda/distribution/lambda-package-type-distribution.html#sana-lambda-package-type-distribution","68":"/sana/lambda/distribution/lambda-package-type-distribution.html#use-cases","69":"/sana/lambda/distribution/lambda-region-distribution.html#sana-lambda-region-distribution","70":"/sana/lambda/distribution/lambda-region-distribution.html#use-cases","71":"/sana/lambda/distribution/lambda-runtime-distribution.html#sana-lambda-runtime-distribution","72":"/sana/lambda/distribution/lambda-runtime-distribution.html#use-cases","73":"/sana/lambda/distribution/lambda-security-group-distribution.html#sana-lambda-security-group-distribution","74":"/sana/lambda/distribution/lambda-security-group-distribution.html#configuration-purpose","75":"/sana/lambda/distribution/lambda-security-group-distribution.html#use-cases","76":"/sana/lambda/distribution/lambda-security-group-distribution.html#tips","77":"/sana/lambda/distribution/lambda-snapstart-status-distribution.html#sana-lambda-snapstart-status-distribution","78":"/sana/lambda/distribution/lambda-snapstart-status-distribution.html#configuration-purpose","79":"/sana/lambda/distribution/lambda-snapstart-status-distribution.html#use-cases","80":"/sana/lambda/distribution/lambda-snapstart-status-distribution.html#tips","81":"/sana/lambda/distribution/lambda-snapstart-status-distribution.html#related-articles","82":"/sana/lambda/distribution/lambda-sys-log-lvl-distribution.html#sana-lambda-sys-log-lvl-distribution","83":"/sana/lambda/distribution/lambda-sys-log-lvl-distribution.html#configuration-purpose","84":"/sana/lambda/distribution/lambda-sys-log-lvl-distribution.html#use-cases","85":"/sana/lambda/distribution/lambda-sys-log-lvl-distribution.html#analyze-logging-kpis","86":"/sana/lambda/distribution/lambda-sys-log-lvl-distribution.html#tips","87":"/sana/lambda/distribution/lambda-tracing-mode-distribution.html#sana-lambda-tracing-mode-distribution","88":"/sana/lambda/distribution/lambda-tracing-mode-distribution.html#use-cases","89":"/sana/lambda/distribution/lambda-vpc-distribution.html#sana-lambda-vpc-distribution","90":"/sana/lambda/distribution/lambda-vpc-distribution.html#use-cases"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[2,1,22],"1":[2,3,22],"2":[3,3,11],"3":[3,1,23],"4":[2,3,19],"5":[5,1,25],"6":[2,5,13],"7":[4,1,26],"8":[2,4,21],"9":[4,1,28],"10":[3,1,16],"11":[2,3,23],"12":[2,3,23],"13":[3,3,31],"14":[2,1,73],"15":[4,3,161],"16":[2,3,18],"17":[2,1,13],"18":[3,2,33],"19":[3,2,59],"20":[3,2,55],"21":[4,2,23],"22":[3,2,22],"23":[1,1,15],"24":[2,1,13],"25":[2,1,27],"26":[4,1,22],"27":[2,1,14],"28":[2,3,22],"29":[3,3,37],"30":[6,1,26],"31":[2,6,30],"32":[4,1,19],"33":[5,1,22],"34":[2,5,19],"35":[4,1,19],"36":[3,1,20],"37":[2,3,20],"38":[5,1,22],"39":[2,5,16],"40":[4,1,64],"41":[2,4,14],"42":[4,1,61],"43":[2,4,12],"44":[6,1,40],"45":[2,6,36],"46":[2,6,18],"47":[3,6,11],"48":[1,6,64],"49":[4,1,22],"50":[2,4,17],"51":[1,4,59],"52":[4,4,11],"53":[6,1,30],"54":[4,1,34],"55":[2,4,22],"56":[2,4,12],"57":[5,1,24],"58":[2,5,26],"59":[5,1,26],"60":[2,5,20],"61":[2,5,21],"62":[3,5,11],"63":[1,5,45],"64":[2,5,8],"65":[4,1,25],"66":[2,4,19],"67":[5,1,22],"68":[2,5,23],"69":[4,1,24],"70":[2,4,22],"71":[4,1,27],"72":[2,4,19],"73":[5,1,29],"74":[2,5,29],"75":[2,5,20],"76":[1,5,62],"77":[5,1,26],"78":[2,5,13],"79":[2,5,22],"80":[1,5,58],"81":[2,5,7],"82":[6,1,35],"83":[2,6,35],"84":[2,6,18],"85":[3,6,11],"86":[1,6,64],"87":[5,1,23],"88":[2,5,17],"89":[4,1,23],"90":[2,4,18]},"averageFieldLength":[2.8901098901098896,2.978021978021978,27.71428571428572],"storedFields":{"0":{"title":"sana ddb","titles":[]},"1":{"title":"Example Usage","titles":["sana ddb",null]},"2":{"title":"KPI-Specific Commands","titles":["sana ddb",null]},"3":{"title":"sana ddb-count","titles":[]},"4":{"title":"Use-cases","titles":["sana ddb-count"]},"5":{"title":"sana ddb-delete-protection-distribution","titles":[]},"6":{"title":"Use-cases","titles":["sana ddb-delete-protection-distribution"]},"7":{"title":"sana ddb-region-distribution","titles":[]},"8":{"title":"Use-cases","titles":["sana ddb-region-distribution"]},"9":{"title":"ddb-table-status-distribution","titles":[]},"10":{"title":"Full Service Analysis","titles":[]},"11":{"title":"Lambda Functions","titles":["Full Service Analysis"]},"12":{"title":"DynamoDB Tables","titles":["Full Service Analysis"]},"13":{"title":"Analyzing Multiple Services","titles":["Full Service Analysis"]},"14":{"title":"Get Started","titles":[]},"15":{"title":"Running your first scan","titles":["Get Started",null]},"16":{"title":"Next Steps","titles":["Get Started",null]},"17":{"title":"Global Flags","titles":[]},"18":{"title":"--profile, -p","titles":["Global Flags"]},"19":{"title":"--region, -r","titles":["Global Flags"]},"20":{"title":"--output, -o","titles":["Global Flags"]},"21":{"title":"--no-progress-bar","titles":["Global Flags"]},"22":{"title":"--silent-mode","titles":["Global Flags"]},"23":{"title":"Installation","titles":[]},"24":{"title":"Pre-requisites","titles":["Installation"]},"25":{"title":"Initial Installation","titles":["Installation"]},"26":{"title":"Upgrading an Existing Installation","titles":["Installation"]},"27":{"title":"sana lambda","titles":[]},"28":{"title":"Example Usage","titles":["sana lambda",null]},"29":{"title":"KPI-Specific Commands","titles":["sana lambda",null]},"30":{"title":"sana lambda-avg-ephemeral-storage-size","titles":[]},"31":{"title":"More Examples","titles":["sana lambda-avg-ephemeral-storage-size"]},"32":{"title":"sana lambda-avg-memory","titles":[]},"33":{"title":"sana lambda-avg-package-size","titles":[]},"34":{"title":"Use-cases","titles":["sana lambda-avg-package-size"]},"35":{"title":"sana lambda-avg-timeout","titles":[]},"36":{"title":"sana lambda-count","titles":[]},"37":{"title":"Use-cases","titles":["sana lambda-count"]},"38":{"title":"sana lambda-max-package-size","titles":[]},"39":{"title":"Use-cases","titles":["sana lambda-max-package-size"]},"40":{"title":"sana lambda-logging-kpis","titles":[]},"41":{"title":"Combined Commands","titles":["sana lambda-logging-kpis"]},"42":{"title":"sana lambda-vpc-kpis","titles":[]},"43":{"title":"Combined Commands","titles":["sana lambda-vpc-kpis"]},"44":{"title":"sana lambda-app-log-lvl-distribution","titles":[]},"45":{"title":"Configuration Purpose","titles":["sana lambda-app-log-lvl-distribution"]},"46":{"title":"Use-cases","titles":["sana lambda-app-log-lvl-distribution"]},"47":{"title":"Analyze Logging KPIs","titles":["sana lambda-app-log-lvl-distribution"]},"48":{"title":"Tips","titles":["sana lambda-app-log-lvl-distribution"]},"49":{"title":"sana lambda-architecture-distribution","titles":[]},"50":{"title":"Use-cases","titles":["sana lambda-architecture-distribution"]},"51":{"title":"Tips","titles":["sana lambda-architecture-distribution"]},"52":{"title":"Related Benchmarks and Materials","titles":["sana lambda-architecture-distribution"]},"53":{"title":"sana lambda-attached-layer-count-distribution","titles":[]},"54":{"title":"sana lambda-dlq-distribution","titles":[]},"55":{"title":"Use-cases","titles":["sana lambda-dlq-distribution"]},"56":{"title":"Notes:","titles":["sana lambda-dlq-distribution"]},"57":{"title":"sana lambda-ephemeral-storage-distribution","titles":[]},"58":{"title":"Use-cases","titles":["sana lambda-ephemeral-storage-distribution"]},"59":{"title":"sana lambda-log-format-distribution","titles":[]},"60":{"title":"Configuration Purpose","titles":["sana lambda-log-format-distribution"]},"61":{"title":"Use-cases","titles":["sana lambda-log-format-distribution"]},"62":{"title":"Analyze Logging KPIs","titles":["sana lambda-log-format-distribution"]},"63":{"title":"Tips","titles":["sana lambda-log-format-distribution"]},"64":{"title":"Related Articles","titles":["sana lambda-log-format-distribution"]},"65":{"title":"sana lambda-memory-distribution","titles":[]},"66":{"title":"Use-cases","titles":["sana lambda-memory-distribution"]},"67":{"title":"sana lambda-package-type-distribution","titles":[]},"68":{"title":"Use-cases","titles":["sana lambda-package-type-distribution"]},"69":{"title":"sana lambda-region-distribution","titles":[]},"70":{"title":"Use-cases","titles":["sana lambda-region-distribution"]},"71":{"title":"sana lambda-runtime-distribution","titles":[]},"72":{"title":"Use-cases","titles":["sana lambda-runtime-distribution"]},"73":{"title":"sana lambda-security-group-distribution","titles":[]},"74":{"title":"Configuration Purpose","titles":["sana lambda-security-group-distribution"]},"75":{"title":"Use-cases","titles":["sana lambda-security-group-distribution"]},"76":{"title":"Tips","titles":["sana lambda-security-group-distribution"]},"77":{"title":"sana lambda-snapstart-status-distribution","titles":[]},"78":{"title":"Configuration Purpose","titles":["sana lambda-snapstart-status-distribution"]},"79":{"title":"Use-cases","titles":["sana lambda-snapstart-status-distribution"]},"80":{"title":"Tips","titles":["sana lambda-snapstart-status-distribution"]},"81":{"title":"Related Articles","titles":["sana lambda-snapstart-status-distribution"]},"82":{"title":"sana lambda-sys-log-lvl-distribution","titles":[]},"83":{"title":"Configuration Purpose","titles":["sana lambda-sys-log-lvl-distribution"]},"84":{"title":"Use-cases","titles":["sana lambda-sys-log-lvl-distribution"]},"85":{"title":"Analyze Logging KPIs","titles":["sana lambda-sys-log-lvl-distribution"]},"86":{"title":"Tips","titles":["sana lambda-sys-log-lvl-distribution"]},"87":{"title":"sana lambda-tracing-mode-distribution","titles":[]},"88":{"title":"Use-cases","titles":["sana lambda-tracing-mode-distribution"]},"89":{"title":"sana lambda-vpc-distribution","titles":[]},"90":{"title":"Use-cases","titles":["sana lambda-vpc-distribution"]}},"dirtCount":0,"index":[["java",{"2":{"79":1,"80":3}}],["java21",{"2":{"63":1}}],["java17",{"2":{"63":1}}],["java11",{"2":{"63":1}}],["java8",{"2":{"63":1}}],["json",{"2":{"15":1,"40":1,"59":2,"61":1,"63":2}}],["query",{"2":{"63":1}}],["queue",{"2":{"55":1}}],["quantifying",{"2":{"61":1}}],["quality",{"2":{"55":1}}],["quickly",{"2":{"80":1}}],["quick",{"2":{"40":1,"42":1}}],["9",{"2":{"44":1,"63":1}}],["92",{"2":{"33":1}}],["94",{"2":{"15":2}}],["helpful",{"2":{"80":1}}],["heavy",{"2":{"51":1}}],["heaviest",{"2":{"38":1,"39":2}}],["hindering",{"2":{"76":1}}],["how",{"2":{"23":1}}],["horizontal",{"2":{"20":1}}],["has",{"2":{"76":1}}],["hand",{"2":{"18":1,"20":1}}],["have",{"2":{"14":1,"51":1}}],["89",{"2":{"15":1,"42":1}}],["88",{"2":{"15":1,"42":1}}],["8",{"2":{"15":1,"35":1,"63":1}}],["zip",{"2":{"15":1,"67":1}}],["60",{"2":{"38":1}}],["6",{"2":{"15":4,"40":1}}],["64",{"2":{"15":1,"49":1,"50":1,"51":3}}],["66",{"2":{"15":5}}],["67",{"2":{"15":7}}],["695",{"2":{"15":1}}],["75",{"2":{"54":1}}],["759",{"2":{"33":1}}],["78",{"2":{"15":3}}],["7",{"2":{"15":1,"63":1}}],["kb",{"2":{"15":1}}],["kpis",{"0":{"40":1,"42":1,"47":1,"62":1,"85":1},"1":{"41":1,"43":1},"2":{"4":1,"29":2,"40":2,"42":3,"47":1,"62":1,"85":1}}],["kpi",{"0":{"2":1,"29":1}}],["☕",{"2":{"15":1,"40":1,"42":1}}],["`",{"2":{"15":2,"40":2,"42":2}}],["|████████████████████████████████████████|",{"2":{"15":1,"40":1,"42":1}}],["|",{"2":{"15":18,"40":18,"42":18}}],["xxxxx",{"2":{"89":1}}],["xxxxxxx3",{"2":{"73":1}}],["xxxxxxx2",{"2":{"73":1}}],["xxxxxxx1",{"2":{"73":1}}],["xxxxxxv3",{"2":{"15":1}}],["xxxxx32",{"2":{"15":1}}],["x86",{"2":{"15":1,"49":1,"50":1,"51":3,"52":1}}],["x",{"2":{"14":1,"15":4,"25":1,"63":3,"71":4,"87":1}}],["vulnerabilities",{"2":{"75":1}}],["vs",{"2":{"52":1,"61":1,"90":1}}],["volumes",{"2":{"48":1,"86":1}}],["volume",{"2":{"48":2,"86":2}}],["verbose",{"2":{"48":1,"86":1}}],["version",{"2":{"14":2,"25":3,"26":2,"78":1}}],["valid",{"2":{"20":1}}],["value",{"2":{"20":1}}],["vpcs",{"2":{"90":1}}],["vpcconfig",{"2":{"74":1}}],["vpc",{"0":{"42":1,"89":1},"1":{"43":1,"90":1},"2":{"15":3,"20":2,"29":2,"42":7,"43":1,"74":1,"89":5}}],["visit",{"2":{"11":1,"12":1}}],["gates",{"2":{"55":1}}],["gain",{"2":{"51":1}}],["group",{"0":{"73":1},"1":{"74":1,"75":1,"76":1},"2":{"29":1,"43":1,"73":2,"74":2,"76":1}}],["groups",{"2":{"15":1,"42":1,"73":1,"76":3}}],["global",{"0":{"17":1},"1":{"18":1,"19":1,"20":1,"21":1,"22":1},"2":{"17":1,"34":1,"37":1}}],["gb",{"2":{"15":2}}],["g",{"2":{"14":1,"25":1,"26":2}}],["generates",{"2":{"75":1}}],["generating",{"2":{"4":1}}],["generally",{"2":{"51":1}}],["getting",{"2":{"22":1}}],["get",{"0":{"14":1},"1":{"15":1,"16":1},"2":{"3":1,"5":1,"9":1,"20":1,"36":1,"40":1,"42":1,"44":1,"49":1,"53":1,"54":1,"57":1,"59":1,"65":1,"67":1,"69":1,"71":1,"73":1,"77":1,"82":1,"87":1,"89":1}}],["list",{"2":{"74":1}}],["like",{"2":{"63":1}}],["line",{"2":{"14":1}}],["ll",{"2":{"48":1,"51":1,"86":1}}],["lvl",{"0":{"44":1,"82":1},"1":{"45":1,"46":1,"47":1,"48":1,"83":1,"84":1,"85":1,"86":1},"2":{"29":2,"41":2,"44":1,"82":1}}],["lowering",{"2":{"80":1}}],["lower",{"2":{"45":1,"51":1,"83":1}}],["long",{"2":{"20":1}}],["logformat",{"2":{"60":1}}],["loggingconfig",{"2":{"45":1,"60":1,"83":1}}],["logging",{"0":{"40":1,"47":1,"62":1,"85":1},"1":{"41":1},"2":{"29":1,"40":3,"44":1,"46":2,"47":2,"61":1,"62":2,"63":1,"82":1,"84":2,"85":2}}],["logs",{"2":{"22":1,"45":2,"48":4,"60":1,"63":2,"64":1,"83":2,"86":4}}],["log",{"0":{"44":1,"59":1,"82":1},"1":{"45":1,"46":1,"47":1,"48":1,"60":1,"61":1,"62":1,"63":1,"64":1,"83":1,"84":1,"85":1,"86":1},"2":{"15":3,"29":3,"40":3,"41":3,"44":2,"45":1,"48":5,"59":3,"61":1,"82":2,"83":1,"86":5}}],["label",{"2":{"56":1}}],["latency",{"2":{"51":1,"79":1,"80":1}}],["latest",{"2":{"14":1,"23":1,"26":1}}],["larger",{"2":{"48":1,"86":1}}],["large",{"2":{"48":1,"58":2,"86":1}}],["layer",{"0":{"53":1},"2":{"29":1,"53":1}}],["layers",{"2":{"15":4,"53":5}}],["lambda",{"0":{"11":1,"27":1,"30":1,"32":1,"33":1,"35":1,"36":1,"38":1,"40":1,"42":1,"44":1,"49":1,"53":1,"54":1,"57":1,"59":1,"65":1,"67":1,"69":1,"71":1,"73":1,"77":1,"82":1,"87":1,"89":1},"1":{"28":1,"29":1,"31":1,"34":1,"37":1,"39":1,"41":1,"43":1,"45":1,"46":1,"47":1,"48":1,"50":1,"51":1,"52":1,"55":1,"56":1,"58":1,"60":1,"61":1,"62":1,"63":1,"64":1,"66":1,"68":1,"70":1,"72":1,"74":1,"75":1,"76":1,"78":1,"79":1,"80":1,"81":1,"83":1,"84":1,"85":1,"86":1,"88":1,"90":1},"2":{"11":5,"13":1,"14":1,"15":59,"18":3,"19":3,"20":5,"21":1,"22":1,"27":3,"28":3,"29":23,"30":2,"31":2,"32":1,"33":2,"35":1,"36":3,"38":2,"40":7,"41":3,"42":9,"43":2,"44":3,"45":2,"47":1,"49":2,"51":1,"52":1,"53":4,"54":3,"57":2,"59":3,"60":1,"62":1,"63":1,"64":1,"65":2,"67":2,"69":2,"70":1,"71":2,"73":6,"74":2,"76":1,"77":3,"80":3,"81":1,"82":3,"83":2,"85":1,"87":2,"89":3,"90":1}}],["letter",{"2":{"55":1}}],["least",{"2":{"76":1}}],["leading",{"2":{"51":1}}],["learn",{"2":{"11":1,"12":1,"16":1}}],["level",{"2":{"10":1,"15":2,"40":2,"44":2,"45":1,"46":3,"48":2,"76":1,"79":2,"82":2,"83":1,"84":3,"86":2}}],["your",{"0":{"15":1},"2":{"14":1,"25":2,"30":1,"45":1,"48":1,"60":1,"68":1,"72":1,"80":1,"83":1,"86":1}}],["you",{"2":{"10":1,"13":1,"14":3,"18":1,"19":3,"20":1,"21":1,"22":1,"25":1,"48":3,"51":3,"76":4,"86":3}}],["03e07f6b",{"2":{"42":1}}],["05",{"2":{"38":1}}],["00",{"2":{"15":4,"54":2,"71":2,"73":3}}],["0s",{"2":{"15":1,"40":1,"42":1}}],["0",{"2":{"9":3,"15":5,"53":1,"77":1}}],["both",{"2":{"76":1}}],["building",{"2":{"51":1}}],["built",{"2":{"24":1,"79":1}}],["bar",{"0":{"21":1},"2":{"21":3}}],["based",{"2":{"5":1,"9":1,"20":2,"48":1,"51":1,"53":1,"55":1,"80":1,"86":1}}],["benchmarks",{"0":{"52":1}}],["behavior",{"2":{"17":1}}],["behaviors",{"2":{"16":1}}],["be",{"2":{"16":1,"72":1,"80":1}}],["below",{"2":{"14":2,"15":2}}],["bytes",{"2":{"15":1,"33":1}}],["by",{"2":{"8":1,"11":1,"12":1,"14":2,"15":16,"18":1,"19":2,"20":2,"40":3,"42":2,"44":2,"45":1,"49":1,"53":1,"54":2,"57":1,"59":2,"65":1,"67":1,"69":1,"70":1,"73":2,"76":1,"77":2,"80":2,"82":2,"83":1,"87":1,"89":2}}],["32",{"2":{"82":1}}],["329",{"2":{"59":1}}],["322",{"2":{"44":1,"82":1}}],["348",{"2":{"36":1}}],["33",{"2":{"15":8}}],["3",{"2":{"7":1,"15":2,"44":1,"53":2}}],["490",{"2":{"69":1}}],["4edf8025",{"2":{"42":1}}],["44",{"2":{"15":5,"82":1}}],["4",{"2":{"7":1,"15":3,"40":1,"54":1}}],["450",{"2":{"87":1}}],["455",{"2":{"59":1}}],["45",{"2":{"3":1,"53":1}}],["500",{"2":{"71":1}}],["50",{"2":{"71":3,"73":2}}],["512",{"2":{"15":2,"30":1,"57":1,"65":1}}],["56",{"2":{"15":11}}],["540",{"2":{"15":1}}],["53",{"2":{"15":1}}],["5",{"2":{"5":1,"15":13,"44":2,"57":1,"67":1}}],["feature",{"2":{"80":2}}],["fallback",{"2":{"70":1}}],["fatal",{"2":{"15":1,"40":1,"44":1}}],["find",{"2":{"76":1}}],["finding",{"2":{"39":1}}],["file",{"2":{"58":1}}],["filtering",{"2":{"45":1,"48":2,"83":1,"86":2}}],["filter",{"2":{"45":1,"83":1}}],["first",{"0":{"15":1}}],["from",{"2":{"15":1,"22":1,"31":2,"40":1,"42":1}}],["focus",{"2":{"48":1,"86":1}}],["foo",{"2":{"18":2,"31":2}}],["found",{"2":{"15":1,"36":1}}],["following",{"2":{"15":1,"25":1,"26":1,"40":1,"41":1,"42":1,"43":1,"63":1}}],["forensics",{"2":{"75":1}}],["form",{"2":{"20":1,"58":1}}],["formatted",{"2":{"20":1}}],["format",{"0":{"59":1},"1":{"60":1,"61":1,"62":1,"63":1,"64":1},"2":{"15":1,"29":1,"40":1,"41":1,"59":3,"60":1,"61":2,"63":1}}],["for",{"2":{"4":1,"6":1,"8":2,"10":1,"14":1,"18":2,"19":1,"20":2,"34":2,"37":2,"39":2,"40":1,"42":1,"45":2,"46":2,"50":2,"55":3,"58":3,"61":2,"63":1,"66":2,"68":1,"70":3,"72":2,"75":4,"78":1,"79":2,"80":5,"83":2,"84":2,"88":2,"90":2}}],["functionality",{"2":{"76":1}}],["function",{"2":{"15":1,"24":1,"33":1,"35":1,"37":2,"38":1,"44":1,"45":1,"49":1,"54":1,"57":1,"59":1,"60":1,"65":1,"67":1,"69":1,"70":1,"71":1,"73":1,"74":2,"76":1,"77":1,"78":1,"80":2,"82":1,"83":1,"87":1,"89":1}}],["functions",{"0":{"11":1},"2":{"14":2,"15":41,"27":1,"30":1,"36":2,"39":2,"40":15,"42":5,"44":7,"46":1,"49":2,"51":1,"53":4,"55":1,"56":1,"57":3,"58":3,"59":2,"61":1,"65":4,"67":2,"69":2,"70":1,"71":4,"72":1,"73":3,"76":2,"77":2,"79":1,"80":2,"82":4,"84":1,"87":2,"88":1,"89":2,"90":1}}],["full",{"0":{"10":1},"1":{"11":1,"12":1,"13":1},"2":{"15":2,"27":1}}],["future",{"2":{"6":1}}],["flags",{"0":{"17":1},"1":{"18":1,"19":1,"20":1,"21":1,"22":1},"2":{"16":1,"17":1}}],["flag",{"2":{"5":1,"18":1,"19":2,"21":2,"22":1}}],["observable",{"2":{"88":1}}],["overly",{"2":{"76":2}}],["overview",{"2":{"40":1,"42":1}}],["option",{"2":{"63":1}}],["optimize",{"2":{"76":1}}],["optimizing",{"2":{"50":1}}],["optimizationstatus",{"2":{"78":1}}],["optimization",{"2":{"34":1,"77":1,"79":1}}],["old",{"2":{"26":1}}],["outside",{"2":{"90":1}}],["outliers",{"2":{"66":1,"68":1}}],["out",{"2":{"20":3}}],["output",{"0":{"20":1},"2":{"20":2,"63":1}}],["o",{"0":{"20":1},"2":{"20":3,"54":1}}],["or",{"2":{"15":1,"18":1,"19":1,"23":1,"46":1,"48":1,"51":1,"59":1,"76":1,"79":1,"84":1,"86":1}}],["order",{"2":{"14":1}}],["orphaned",{"2":{"8":1}}],["only",{"2":{"45":1,"83":1}}],["one",{"2":{"26":1}}],["on",{"2":{"5":1,"9":1,"15":1,"24":1,"48":2,"51":2,"53":1,"77":1,"79":1,"86":2}}],["often",{"2":{"21":1}}],["off",{"2":{"15":1,"77":1}}],["of",{"2":{"0":1,"3":1,"4":1,"6":2,"8":1,"13":1,"14":3,"15":2,"16":1,"19":1,"20":1,"23":1,"24":1,"25":1,"26":1,"27":1,"34":1,"36":2,"37":1,"39":2,"40":1,"42":1,"48":3,"50":1,"51":1,"53":2,"55":2,"58":5,"61":2,"66":3,"70":3,"72":2,"74":1,"76":2,"79":2,"80":3,"86":3,"88":1,"90":1}}],["python3",{"2":{"63":6}}],["posing",{"2":{"76":1}}],["post",{"2":{"4":1,"37":1}}],["potential",{"2":{"75":1,"76":1}}],["potentially",{"2":{"48":1,"86":1}}],["power",{"2":{"51":1}}],["policies",{"2":{"46":1,"76":1,"84":1,"88":1}}],["purpose",{"0":{"45":1,"60":1,"74":1,"78":1,"83":1}}],["p",{"0":{"18":1},"2":{"18":2,"31":1}}],["packaging",{"2":{"67":1,"68":2}}],["package",{"0":{"33":1,"38":1,"67":1},"1":{"34":1,"39":1,"68":1},"2":{"14":1,"15":3,"22":1,"29":3,"33":3,"34":2,"38":3,"39":1,"67":1}}],["pay",{"2":{"48":1,"86":1}}],["passing",{"2":{"20":2}}],["pass",{"2":{"19":1}}],["passthrough",{"2":{"15":1,"87":1}}],["page",{"2":{"11":1,"12":1,"17":1,"23":1}}],["particular",{"2":{"10":1}}],["percentile",{"2":{"90":1}}],["permissive",{"2":{"76":1}}],["performing",{"2":{"75":1}}],["performance",{"2":{"51":1,"52":1,"76":1,"80":2,"81":1}}],["per",{"2":{"7":1}}],["privilege",{"2":{"76":1}}],["principle",{"2":{"76":1}}],["prints",{"2":{"20":3}}],["printed",{"2":{"20":1}}],["profiling",{"2":{"90":1}}],["profiles",{"2":{"18":2}}],["profile",{"0":{"18":1},"2":{"15":2,"18":5,"31":2,"40":1,"42":1}}],["property",{"2":{"45":2,"60":1,"74":1,"78":1,"83":2}}],["progress",{"0":{"21":1},"2":{"21":3,"50":1,"55":1,"61":1,"72":1}}],["produce",{"2":{"15":1,"19":1}}],["provide",{"2":{"14":1,"19":1}}],["protections",{"2":{"14":1}}],["protection",{"0":{"5":1},"1":{"6":1},"2":{"2":1,"5":4}}],["prevent",{"2":{"22":1}}],["previous",{"2":{"20":1}}],["pre",{"0":{"24":1},"2":{"4":1,"37":1}}],["might",{"2":{"76":1}}],["migrations",{"2":{"61":1}}],["migration",{"2":{"50":1,"72":1}}],["misconfigurations",{"2":{"75":1}}],["misconfigured",{"2":{"8":1,"58":1}}],["modify",{"2":{"16":1}}],["mode",{"0":{"22":1,"87":1},"1":{"88":1},"2":{"15":1,"20":1,"22":3,"29":1,"87":2}}],["most",{"2":{"16":1,"66":1,"68":1}}],["more",{"0":{"31":1},"2":{"11":1,"12":1,"14":1,"16":1,"48":1,"51":2,"63":1,"86":1}}],["my",{"2":{"15":1,"54":1}}],["mb",{"2":{"15":4,"30":1,"32":1,"38":1}}],["materials",{"0":{"52":1}}],["make",{"2":{"48":1,"63":1,"86":1}}],["managing",{"2":{"48":1,"86":1}}],["many",{"2":{"14":1}}],["machine",{"2":{"18":1,"25":1}}],["max",{"0":{"38":1},"1":{"39":1},"2":{"15":1,"29":1,"38":3}}],["main",{"2":{"8":1,"70":1}}],["messages",{"2":{"48":1,"86":1}}],["means",{"2":{"45":1,"83":1}}],["measurement",{"2":{"39":1}}],["measuring",{"2":{"4":1,"34":1,"37":1,"46":1,"50":1,"55":1,"58":1,"61":1,"70":1,"72":1,"79":1,"84":1,"88":1}}],["memory",{"0":{"32":1,"65":1},"1":{"66":1},"2":{"15":2,"29":2,"32":3,"51":1,"65":2,"66":2}}],["metrics",{"2":{"14":1,"34":1,"40":1,"42":1}}],["multiple",{"0":{"13":1},"2":{"13":1}}],["multi",{"2":{"1":1,"19":1,"28":1}}],["currently",{"2":{"63":1}}],["ci",{"2":{"55":1}}],["cisc",{"2":{"51":1}}],["cloudwatch",{"2":{"45":1,"48":1,"60":1,"64":1,"83":1,"86":1}}],["cli",{"2":{"14":1,"15":1,"18":2,"20":5,"23":1}}],["critical",{"2":{"22":1}}],["credentials",{"2":{"9":1}}],["creating",{"2":{"9":1,"80":1}}],["cheaper",{"2":{"50":1}}],["checks",{"2":{"75":1}}],["check",{"2":{"14":1,"25":1}}],["chips",{"2":{"50":1}}],["changes",{"2":{"80":1}}],["change",{"2":{"61":1}}],["challenging",{"2":{"48":1,"86":1}}],["chart",{"2":{"20":4,"54":1}}],["capture",{"2":{"48":1,"86":1}}],["can",{"2":{"10":1,"14":1,"18":1,"19":2,"21":1,"22":1,"25":1,"48":3,"63":1,"76":3,"80":2,"86":3}}],["cases",{"0":{"4":1,"6":1,"8":1,"34":1,"37":1,"39":1,"46":1,"50":1,"55":1,"58":1,"61":1,"66":1,"68":1,"70":1,"72":1,"75":1,"79":1,"84":1,"88":1,"90":1}}],["code",{"2":{"80":1}}],["cost",{"2":{"50":1,"51":1,"52":1,"63":1,"80":1}}],["costs",{"2":{"48":1,"51":1,"86":1}}],["cold",{"2":{"34":1,"80":1}}],["conduct",{"2":{"76":1}}],["concurrency",{"2":{"70":1}}],["contains",{"2":{"70":1}}],["controls",{"2":{"20":1}}],["configured",{"2":{"56":1}}],["configurations",{"2":{"66":1}}],["configuration",{"0":{"45":1,"60":1,"74":1,"78":1,"83":1},"2":{"44":1,"45":1,"66":1,"82":1,"83":1}}],["constrained",{"2":{"51":1}}],["console",{"2":{"20":1}}],["could",{"2":{"16":1}}],["counterparts",{"2":{"51":1}}],["count",{"0":{"3":1,"36":1,"53":1},"1":{"4":1,"37":1},"2":{"2":1,"3":2,"7":1,"21":1,"29":2,"36":1,"37":2,"53":1}}],["combines",{"2":{"41":1,"43":1}}],["combined",{"0":{"41":1,"43":1}}],["compromised",{"2":{"58":1}}],["computing",{"2":{"51":2}}],["comply",{"2":{"68":1}}],["complex",{"2":{"51":1}}],["completes",{"2":{"14":1}}],["compliance",{"2":{"46":1,"75":1,"84":1,"88":1,"90":1}}],["comparing",{"2":{"37":1,"52":1}}],["commonly",{"2":{"66":1}}],["common",{"2":{"16":1,"68":1}}],["comma",{"2":{"13":1}}],["commands",{"0":{"2":1,"29":1,"41":1,"43":1},"2":{"10":1,"11":1,"12":1,"20":3,"41":1,"43":1,"47":1,"62":1,"85":1}}],["command",{"2":{"0":1,"13":1,"14":3,"15":2,"20":1,"25":2,"26":1,"40":2,"42":2,"68":2}}],[">",{"2":{"3":1,"5":2,"7":3,"9":5,"14":1,"25":1,"30":1,"32":1,"33":1,"35":1,"36":1,"38":1,"42":17,"44":10,"49":2,"53":5,"54":3,"57":3,"59":5,"65":4,"67":2,"69":2,"71":4,"73":4,"77":5,"82":7,"87":2,"89":3}}],["named",{"2":{"18":3,"31":1}}],["new",{"2":{"80":1}}],["network",{"2":{"76":2}}],["necessary",{"2":{"74":1}}],["next",{"0":{"16":1}}],["needs",{"2":{"74":1}}],["need",{"2":{"14":1,"48":1,"86":1}}],["null",{"2":{"15":2,"40":2,"44":1,"45":1,"82":1,"83":1}}],["numbers",{"2":{"58":1}}],["number",{"2":{"3":1,"15":1,"36":1,"53":2}}],["ninja",{"2":{"14":2,"25":2,"26":2}}],["npm",{"2":{"14":2,"25":1,"26":2}}],["non",{"2":{"22":1,"88":1}}],["not",{"2":{"20":1}}],["notes",{"0":{"56":1}}],["note",{"2":{"0":1}}],["nodejs",{"2":{"24":1}}],["nodejs20",{"2":{"15":1,"63":1,"71":1}}],["nodejs18",{"2":{"15":1,"63":1,"71":1}}],["nodejs16",{"2":{"15":1,"63":1,"71":1}}],["nodejs14",{"2":{"15":1,"71":1}}],["no",{"0":{"21":1},"2":{"14":1,"15":2,"21":2,"42":1,"45":1,"54":1,"55":1,"56":2,"80":2,"83":1,"89":1}}],["work",{"2":{"58":1}}],["workloads",{"2":{"51":1,"79":1}}],["well",{"2":{"20":1}}],["west",{"2":{"1":1,"7":1,"28":1,"31":1}}],["warn",{"2":{"15":2,"40":2,"44":1,"82":1}}],["want",{"2":{"13":1}}],["whether",{"2":{"78":1}}],["when",{"2":{"4":1,"74":1}}],["while",{"2":{"51":1}}],["which",{"2":{"45":1,"51":1,"60":1,"76":2,"80":1,"83":1}}],["what",{"2":{"14":1}}],["will",{"2":{"48":1,"86":1}}],["wish",{"2":{"19":1}}],["within",{"2":{"48":1,"86":1}}],["without",{"2":{"48":2,"86":2}}],["with",{"2":{"6":1,"14":1,"51":1,"55":1,"56":1,"58":1,"64":1,"68":1,"74":1,"76":1,"80":1,"81":1}}],["wide",{"2":{"1":1,"19":1,"28":1}}],["280",{"2":{"87":1}}],["24",{"2":{"65":1,"82":1}}],["200",{"2":{"57":1}}],["20",{"2":{"49":1}}],["254",{"2":{"67":1}}],["256",{"2":{"65":1}}],["25",{"2":{"35":1,"54":1,"71":1,"73":2}}],["250",{"2":{"5":1,"71":2}}],["220",{"2":{"49":1}}],["22",{"2":{"15":6}}],["27",{"2":{"15":2}}],["21",{"2":{"7":1,"44":1}}],["2",{"2":{"1":2,"15":4,"19":2,"28":2,"31":1,"42":2,"54":1}}],["each",{"2":{"74":1,"76":1}}],["easier",{"2":{"63":1}}],["east",{"2":{"1":2,"7":1,"15":3,"19":4,"28":2,"31":3,"54":1,"69":1}}],["effective",{"2":{"63":1}}],["efficient",{"2":{"51":2}}],["esbuild",{"2":{"34":1}}],["execution",{"2":{"80":1}}],["extra",{"2":{"80":1}}],["extremely",{"2":{"80":1}}],["existence",{"2":{"66":1}}],["existing",{"0":{"26":1}}],["examples",{"0":{"31":1}}],["example",{"0":{"1":1,"28":1}}],["equal",{"2":{"20":1}}],["errors",{"2":{"48":1,"86":1}}],["error",{"2":{"15":1,"40":1,"44":1}}],["ephemeral",{"0":{"30":1,"57":1},"1":{"31":1,"58":1},"2":{"15":2,"29":2,"30":3,"31":2,"57":2,"58":1}}],["eta",{"2":{"15":1,"40":1,"42":1}}],["etc",{"2":{"13":1}}],["eg",{"2":{"13":1}}],["environment",{"2":{"80":1}}],["enhance",{"2":{"76":1}}],["ensure",{"2":{"76":1}}],["energy",{"2":{"51":2}}],["encryption",{"2":{"9":1}}],["engineers",{"2":{"8":1}}],["enabled",{"2":{"5":1,"15":1,"19":1}}],["15",{"2":{"65":1,"89":1}}],["11",{"2":{"15":4,"42":4,"63":1}}],["16",{"2":{"15":2,"42":1,"44":1,"54":2}}],["1205",{"2":{"69":1}}],["128",{"2":{"15":1,"65":1}}],["12",{"2":{"15":5,"40":3,"54":1,"63":1,"71":2}}],["184",{"2":{"32":1}}],["18",{"2":{"15":5,"77":1}}],["177",{"2":{"15":1}}],["17",{"2":{"15":4,"40":2,"42":2}}],["10x",{"2":{"80":1}}],["10",{"2":{"63":1,"65":1,"73":2}}],["10240",{"2":{"57":1,"65":1}}],["1024",{"2":{"15":1,"57":1}}],["1000",{"2":{"71":1}}],["100",{"2":{"15":5,"40":1,"42":1}}],["10500",{"2":{"9":1}}],["1",{"2":{"1":2,"7":3,"9":1,"14":2,"15":17,"19":5,"25":2,"28":2,"31":5,"36":1,"40":8,"53":2,"57":1,"65":1,"69":2,"89":1}}],["risc",{"2":{"51":1}}],["risks",{"2":{"76":1}}],["risk",{"2":{"6":1,"75":1}}],["ray",{"2":{"87":1}}],["ratio",{"2":{"70":1}}],["rates",{"2":{"46":1,"79":1,"84":1}}],["raising",{"2":{"66":1}}],["raw",{"2":{"20":2}}],["rollout",{"2":{"61":1}}],["rollback",{"2":{"4":1,"37":1}}],["role",{"2":{"15":1}}],["related",{"0":{"52":1,"64":1,"81":1},"2":{"40":1,"42":2,"47":1,"62":1,"85":1,"90":1}}],["release",{"2":{"23":1}}],["refactoring",{"2":{"39":1}}],["retrieve",{"2":{"30":1,"32":1,"33":1,"35":1,"38":1}}],["requires",{"2":{"24":1}}],["required",{"2":{"19":1}}],["requisites",{"0":{"24":1}}],["reduced",{"2":{"51":1}}],["reduce",{"2":{"19":1}}],["reporting",{"2":{"90":1}}],["report",{"2":{"19":2}}],["reports",{"2":{"15":1}}],["regardless",{"2":{"51":1}}],["regular",{"2":{"18":1}}],["regions",{"2":{"8":1,"15":2,"19":2,"40":1,"42":1,"70":3}}],["region",{"0":{"7":1,"19":1,"69":1},"1":{"8":1,"70":1},"2":{"1":1,"2":1,"7":2,"8":1,"15":1,"19":6,"20":1,"28":1,"29":1,"40":1,"42":1,"46":1,"69":2,"79":1,"84":1}}],["regional",{"2":{"1":1,"28":1,"34":1}}],["restored",{"2":{"80":1}}],["restrictive",{"2":{"76":1}}],["response",{"2":{"75":1}}],["result",{"2":{"48":1,"86":1}}],["results",{"2":{"20":1}}],["resources",{"2":{"14":1,"31":2,"74":1}}],["residing",{"2":{"8":1}}],["r",{"0":{"19":1},"2":{"1":2,"19":3,"28":2,"31":2}}],["running",{"0":{"15":1},"2":{"14":1,"15":2,"40":1,"42":1,"90":1}}],["runtimes",{"2":{"63":1}}],["runtime",{"0":{"71":1},"1":{"72":1},"2":{"14":1,"15":1,"20":2,"29":1,"51":1,"71":2,"72":1}}],["runs",{"2":{"1":3,"19":2,"28":3}}],["run",{"2":{"0":1,"10":1,"11":1,"12":1,"15":1,"25":1,"26":1,"27":1,"47":1,"62":1,"85":1}}],["$",{"2":{"1":3,"18":3,"19":3,"20":5,"26":2,"28":3,"31":2}}],["uninstall",{"2":{"26":2}}],["understanding",{"2":{"76":1}}],["underlying",{"2":{"49":1}}],["under",{"2":{"0":1,"56":1}}],["up",{"2":{"80":1}}],["upgrading",{"0":{"26":1}}],["update",{"2":{"23":1}}],["updating",{"2":{"9":1}}],["utilizes",{"2":{"18":1,"58":1}}],["using",{"2":{"13":1,"14":2,"15":1,"18":1,"25":1,"31":2,"40":1,"42":1,"64":1}}],["uses",{"2":{"15":1,"18":1,"61":1}}],["useful",{"2":{"4":2,"6":1,"8":2,"34":2,"37":2,"46":2,"50":2,"55":3,"58":3,"61":2,"66":2,"70":3,"72":2,"75":3,"79":2,"84":2,"88":2,"90":2}}],["use",{"0":{"4":1,"6":1,"8":1,"34":1,"37":1,"39":1,"46":1,"50":1,"55":1,"58":1,"61":1,"66":1,"68":1,"70":1,"72":1,"75":1,"79":1,"84":1,"88":1,"90":1},"2":{"10":1,"13":1,"14":1,"18":1,"19":1,"21":1,"22":1,"40":1,"42":1,"47":1,"62":1,"68":2,"76":1,"85":1}}],["used",{"2":{"0":1,"3":1,"5":1,"7":1,"8":2,"9":1,"11":1,"12":1,"14":2,"16":1,"17":1,"21":1,"27":1,"30":1,"32":1,"33":1,"35":1,"36":1,"38":1,"39":2,"44":1,"45":1,"49":1,"51":1,"53":1,"54":1,"57":1,"59":1,"65":1,"66":1,"67":1,"68":1,"69":1,"70":1,"71":1,"73":1,"77":1,"82":1,"87":1,"89":1}}],["us",{"2":{"1":3,"7":2,"15":3,"19":4,"28":3,"31":4,"54":1,"69":1}}],["usage",{"0":{"1":1,"28":1}}],["id",{"2":{"89":1}}],["ids",{"2":{"74":1}}],["identify",{"2":{"48":1,"68":1,"86":1}}],["identification",{"2":{"6":1,"58":2,"66":1,"70":2,"72":1,"75":1,"88":1}}],["improving",{"2":{"81":1}}],["improve",{"2":{"80":2}}],["impact",{"2":{"4":1,"34":1,"37":1,"39":1,"48":1,"86":1}}],["image",{"2":{"67":1}}],["iam",{"2":{"15":1}}],["i",{"2":{"14":1,"25":1,"26":1}}],["it",{"2":{"14":1,"24":1,"48":1,"63":1,"74":1,"80":1,"86":1}}],["if",{"2":{"13":1,"14":1,"19":1,"51":1}}],["invocations",{"2":{"80":1}}],["inside",{"2":{"74":1,"90":1}}],["instruction",{"2":{"51":2}}],["instructs",{"2":{"15":1}}],["installing",{"2":{"55":1}}],["installed",{"2":{"25":1}}],["installation",{"0":{"23":1,"25":1,"26":1},"1":{"24":1,"25":1,"26":1},"2":{"14":1,"55":1}}],["install",{"2":{"14":1,"23":1,"25":1,"26":1}}],["initiatives",{"2":{"72":1}}],["initialized",{"2":{"80":1}}],["initial",{"0":{"25":1}}],["indicates",{"2":{"60":1,"78":1}}],["ingested",{"2":{"48":1,"86":1}}],["incident",{"2":{"75":1}}],["increase",{"2":{"48":1,"86":1}}],["including",{"2":{"48":1,"86":1}}],["inputs",{"2":{"20":1}}],["interact",{"2":{"74":1}}],["interest",{"2":{"19":1}}],["interface",{"2":{"14":1}}],["information",{"2":{"48":1,"76":1,"86":1}}],["info",{"2":{"15":2,"40":2,"44":1,"82":1}}],["inaccessible",{"2":{"9":1}}],["in",{"2":{"0":1,"3":1,"8":2,"13":1,"14":1,"15":3,"19":1,"20":3,"21":1,"25":1,"39":1,"40":1,"42":1,"48":2,"51":1,"60":1,"66":1,"68":1,"70":1,"86":2,"88":1}}],["issues",{"2":{"48":1,"86":1}}],["is",{"2":{"0":2,"3":1,"5":1,"7":1,"9":1,"11":1,"12":1,"14":3,"20":2,"21":1,"24":1,"27":1,"30":1,"32":1,"33":1,"35":1,"36":1,"38":1,"44":1,"45":3,"49":1,"51":2,"53":1,"54":1,"57":1,"59":1,"65":1,"67":1,"69":1,"71":1,"73":1,"74":2,"77":1,"78":1,"80":2,"82":1,"83":2,"87":1,"89":1}}],["typically",{"2":{"51":1,"80":1}}],["type",{"0":{"67":1},"1":{"68":1},"2":{"15":1,"29":1,"67":2,"68":1}}],["tips",{"0":{"48":1,"51":1,"63":1,"76":1,"80":1,"86":1}}],["tip",{"2":{"21":1}}],["times",{"2":{"80":1}}],["time",{"2":{"19":1}}],["timeout",{"0":{"35":1},"2":{"15":1,"29":1,"35":3}}],["temporary",{"2":{"58":1}}],["terms",{"2":{"39":1,"66":1}}],["terminal",{"2":{"20":1}}],["text",{"2":{"15":1,"20":4,"40":1,"59":2,"61":1}}],["target",{"2":{"19":1,"40":1,"42":1}}],["table",{"0":{"9":1},"2":{"2":1,"3":1,"5":1,"7":1,"9":8,"20":5}}],["tables",{"0":{"12":1},"2":{"0":1,"3":1,"5":2,"6":1,"7":3,"8":1,"14":1}}],["thorough",{"2":{"76":1}}],["than",{"2":{"50":1,"51":2}}],["that",{"2":{"16":1,"45":2,"58":2,"61":1,"68":1,"70":1,"72":1,"75":1,"76":2,"83":2}}],["through",{"2":{"48":1,"58":1,"86":1}}],["this",{"2":{"17":1,"19":1,"20":1,"21":1,"23":1,"40":1,"41":1,"42":1,"43":1,"45":1,"48":1,"68":2,"76":1,"80":2,"83":1,"86":1}}],["then",{"2":{"80":1}}],["these",{"2":{"76":1}}],["their",{"2":{"5":1,"9":1,"51":1,"76":1}}],["the",{"2":{"3":1,"4":1,"7":1,"8":1,"9":1,"10":1,"11":1,"12":1,"14":7,"15":9,"16":2,"17":1,"18":1,"19":4,"20":4,"21":2,"22":1,"23":1,"25":4,"26":3,"30":1,"31":2,"32":1,"33":1,"34":1,"35":1,"36":1,"37":1,"38":1,"39":2,"40":3,"41":1,"42":3,"43":1,"44":1,"45":3,"47":1,"48":2,"49":1,"51":2,"53":2,"54":1,"56":1,"57":1,"58":2,"59":1,"60":1,"61":1,"62":1,"63":2,"65":1,"66":1,"67":1,"68":1,"69":1,"71":1,"72":1,"73":1,"74":1,"76":3,"77":1,"78":1,"80":2,"82":1,"83":3,"85":1,"86":2,"87":1,"88":1,"89":1}}],["tracking",{"2":{"55":1}}],["tracing",{"0":{"87":1},"1":{"88":1},"2":{"15":1,"29":1,"87":2,"88":1}}],["trace",{"2":{"15":1,"40":1,"44":1}}],["top",{"2":{"24":1,"79":1}}],["total",{"2":{"3":1,"36":2}}],["to",{"2":{"0":1,"3":1,"5":1,"7":1,"9":1,"10":1,"11":2,"12":2,"13":2,"14":4,"15":3,"16":1,"17":1,"18":1,"19":4,"20":1,"21":1,"22":1,"23":1,"24":1,"25":1,"26":1,"27":1,"30":1,"32":1,"33":1,"34":1,"35":1,"36":1,"37":1,"38":1,"39":1,"40":1,"42":1,"44":1,"45":2,"46":1,"47":1,"48":2,"49":1,"51":1,"53":1,"54":1,"57":1,"59":1,"60":1,"62":1,"63":2,"65":1,"67":1,"68":2,"69":1,"71":1,"72":1,"73":1,"74":3,"76":5,"77":1,"80":3,"82":1,"83":2,"84":1,"85":1,"86":2,"87":1,"88":1,"89":1}}],["data",{"2":{"48":2,"86":2}}],["done",{"2":{"15":1,"40":1,"42":1,"44":1,"59":1,"77":1,"82":1}}],["dlq",{"0":{"54":1},"1":{"55":1,"56":1},"2":{"15":3,"29":1,"54":5,"55":1,"56":2}}],["displayed",{"2":{"22":1,"56":1}}],["disable",{"2":{"21":1}}],["disabled",{"2":{"5":1}}],["distribution",{"0":{"5":1,"7":1,"9":1,"44":1,"49":1,"53":1,"54":1,"57":1,"59":1,"65":1,"67":1,"69":1,"71":1,"73":1,"77":1,"82":1,"87":1,"89":1},"1":{"6":1,"8":1,"45":1,"46":1,"47":1,"48":1,"50":1,"51":1,"52":1,"55":1,"56":1,"58":1,"60":1,"61":1,"62":1,"63":1,"64":1,"66":1,"68":1,"70":1,"72":1,"74":1,"75":1,"76":1,"78":1,"79":1,"80":1,"81":1,"83":1,"84":1,"85":1,"86":1,"88":1,"90":1},"2":{"2":4,"5":2,"7":1,"9":2,"14":1,"15":15,"20":7,"29":16,"40":3,"41":3,"42":2,"43":2,"44":3,"49":2,"53":3,"54":3,"57":2,"59":3,"65":2,"67":2,"69":2,"71":2,"73":3,"76":1,"77":3,"82":3,"87":2,"89":3}}],["during",{"2":{"4":1,"37":1}}],["designed",{"2":{"80":1}}],["describes",{"2":{"17":1,"23":1}}],["deprecated",{"2":{"72":1}}],["deployment",{"2":{"4":2,"37":2}}],["deployments",{"2":{"4":1,"37":1}}],["dead",{"2":{"55":1}}],["debugging",{"2":{"48":1,"86":1}}],["debug",{"2":{"15":2,"40":2,"44":1,"82":1}}],["default",{"2":{"15":3,"16":1,"18":3,"19":1,"20":1,"31":1,"40":1,"42":1,"45":2,"83":2}}],["dedicated",{"2":{"11":1,"12":1}}],["deleting",{"2":{"9":1}}],["deletion",{"2":{"6":1}}],["delete",{"0":{"5":1},"1":{"6":1},"2":{"2":1,"5":4,"14":1}}],["development",{"2":{"0":1}}],["dynamodb",{"0":{"12":1},"2":{"0":1,"3":3,"4":1,"5":1,"6":1,"7":1,"8":1,"9":1,"12":3,"14":1}}],["ddb",{"0":{"0":1,"3":1,"5":1,"7":1,"9":1},"1":{"1":1,"2":1,"4":1,"6":1,"8":1},"2":{"0":3,"1":3,"2":4,"3":1,"5":1,"7":1,"9":1,"12":2,"13":1}}],["adjusting",{"2":{"76":1}}],["adhering",{"2":{"76":1}}],["adoption",{"2":{"46":1,"79":1,"84":1}}],["audits",{"2":{"75":1,"76":1}}],["automation",{"2":{"15":1,"21":1}}],["awareness",{"2":{"66":1}}],["aws",{"2":{"0":1,"3":1,"14":2,"15":4,"18":2,"19":1,"31":1,"48":1,"52":1,"54":1,"64":1,"69":1,"74":1,"80":1,"86":1}}],["amazon",{"2":{"64":1}}],["amount",{"2":{"48":1,"55":1,"58":1,"61":1,"86":1}}],["attaching",{"2":{"74":1}}],["attached",{"0":{"53":1},"2":{"15":2,"29":1,"53":3,"74":1}}],["at",{"2":{"45":1,"80":1,"83":1}}],["against",{"2":{"19":1}}],["aggregation",{"2":{"2":1,"29":1}}],["al2",{"2":{"63":1}}],["alias",{"2":{"41":1,"43":1}}],["aliases",{"2":{"29":1}}],["also",{"2":{"19":1}}],["alter",{"2":{"17":1}}],["allows",{"2":{"76":1}}],["all",{"2":{"0":2,"10":1,"11":1,"12":1,"15":1,"19":1,"30":1,"47":1,"48":1,"62":1,"85":1,"86":1}}],["articles",{"0":{"64":1,"81":1}}],["are",{"2":{"50":1,"51":2,"56":1,"72":1,"76":2}}],["arm",{"2":{"50":1,"52":1}}],["arm64",{"2":{"15":1,"49":1,"51":2}}],["arn",{"2":{"15":1,"54":2}}],["architecture",{"0":{"49":1},"1":{"50":1,"51":1,"52":1},"2":{"15":1,"29":1,"49":2,"51":2}}],["avg",{"0":{"30":1,"32":1,"33":1,"35":1},"1":{"31":1,"34":1},"2":{"15":4,"22":1,"29":4,"30":2,"31":2,"32":2,"33":2,"35":2}}],["average",{"2":{"14":1,"30":1,"32":1,"33":1,"35":1}}],["after",{"2":{"14":1}}],["assigned",{"2":{"76":1}}],["assessments",{"2":{"75":1}}],["associated",{"2":{"6":1,"76":1}}],["as",{"2":{"14":1,"19":1,"48":2,"86":2}}],["about",{"2":{"11":1,"12":1,"16":1,"66":1}}],["achieves",{"2":{"80":1}}],["activated",{"2":{"78":1}}],["activities",{"2":{"34":1,"55":1}}],["active",{"2":{"9":1,"87":1}}],["across",{"2":{"30":1,"46":1,"84":1}}],["access",{"2":{"76":2}}],["accidental",{"2":{"6":1}}],["account",{"2":{"0":1,"1":1,"3":1,"8":2,"14":1,"15":4,"19":2,"28":1,"40":2,"42":2,"46":1,"58":1,"68":1,"70":2,"79":1,"84":1,"88":1}}],["appropriate",{"2":{"76":1}}],["applied",{"2":{"45":1,"83":1}}],["applications",{"2":{"80":1}}],["applicationloglevel",{"2":{"45":1}}],["application",{"2":{"15":1,"40":1,"44":2,"45":2,"46":1,"60":1,"63":1}}],["app",{"0":{"44":1},"1":{"45":1,"46":1,"47":1,"48":1},"2":{"15":1,"29":1,"41":1,"44":1,"54":1}}],["ap",{"2":{"1":1,"7":1,"15":1,"19":3,"28":1,"31":2,"69":1}}],["a",{"2":{"1":2,"10":1,"13":2,"14":1,"15":1,"18":2,"19":2,"20":4,"27":1,"28":2,"36":1,"40":1,"42":1,"48":2,"51":1,"58":1,"74":2,"80":3,"86":2}}],["analyzing",{"0":{"13":1},"2":{"14":1,"42":1,"46":1,"48":1,"76":1,"79":1,"84":1,"86":1}}],["analyze",{"0":{"47":1,"62":1,"85":1},"2":{"7":1,"63":1}}],["analysis",{"0":{"10":1},"1":{"11":1,"12":1,"13":1},"2":{"0":1,"1":3,"10":1,"11":1,"12":1,"15":2,"19":2,"27":1,"28":3,"29":1,"52":1}}],["and",{"0":{"52":1},"2":{"4":1,"14":1,"24":1,"26":1,"31":1,"34":1,"37":1,"45":1,"48":3,"51":1,"52":1,"60":1,"63":2,"70":1,"75":3,"76":1,"83":1,"86":3,"90":1}}],["an",{"0":{"26":1},"2":{"0":1,"1":1,"3":1,"8":2,"14":1,"15":2,"19":3,"28":1,"68":1,"70":2,"74":1,"88":1}}],["snapshot",{"2":{"80":1}}],["snapstart",{"0":{"77":1},"1":{"78":1,"79":1,"80":1,"81":1},"2":{"15":1,"29":1,"77":3,"78":2,"79":1,"80":2,"81":1}}],["some",{"2":{"76":1}}],["soon",{"2":{"72":1}}],["southeast",{"2":{"1":1,"7":1,"15":1,"19":3,"28":1,"31":2,"69":1}}],["specified",{"2":{"78":1}}],["specific",{"0":{"2":1,"29":1},"2":{"11":2,"12":2,"48":1,"86":1}}],["specify",{"2":{"18":1}}],["sys",{"0":{"82":1},"1":{"83":1,"84":1,"85":1,"86":1},"2":{"29":1,"41":1,"82":1}}],["systemloglevel",{"2":{"83":1}}],["systems",{"2":{"50":1,"58":1}}],["system",{"2":{"15":1,"40":1,"60":1,"68":1,"82":2,"83":2,"84":1}}],["syntax",{"2":{"13":1,"18":2,"19":1,"20":3}}],["sg",{"2":{"15":1,"42":1,"73":3}}],["sqs",{"2":{"15":1,"54":1}}],["saving",{"2":{"51":1}}],["san",{"2":{"31":1}}],["sana",{"0":{"0":1,"3":1,"5":1,"7":1,"27":1,"30":1,"32":1,"33":1,"35":1,"36":1,"38":1,"40":1,"42":1,"44":1,"49":1,"53":1,"54":1,"57":1,"59":1,"65":1,"67":1,"69":1,"71":1,"73":1,"77":1,"82":1,"87":1,"89":1},"1":{"1":1,"2":1,"4":1,"6":1,"8":1,"28":1,"29":1,"31":1,"34":1,"37":1,"39":1,"41":1,"43":1,"45":1,"46":1,"47":1,"48":1,"50":1,"51":1,"52":1,"55":1,"56":1,"58":1,"60":1,"61":1,"62":1,"63":1,"64":1,"66":1,"68":1,"70":1,"72":1,"74":1,"75":1,"76":1,"78":1,"79":1,"80":1,"81":1,"83":1,"84":1,"85":1,"86":1,"88":1,"90":1},"2":{"0":2,"1":3,"3":1,"5":1,"7":1,"9":1,"11":3,"12":3,"13":2,"14":6,"15":3,"16":1,"17":1,"18":4,"19":5,"20":5,"21":1,"22":1,"23":1,"24":1,"25":5,"26":4,"27":2,"28":3,"30":1,"31":2,"32":1,"33":1,"35":1,"36":1,"38":1,"40":1,"42":1,"44":1,"49":1,"53":1,"54":1,"57":1,"59":1,"65":1,"67":1,"69":1,"71":1,"73":1,"77":1,"82":1,"87":1,"89":1}}],["sample",{"2":{"15":1}}],["simpler",{"2":{"51":1}}],["sift",{"2":{"48":1,"86":1}}],["significantly",{"2":{"80":1}}],["significant",{"2":{"48":1,"86":1}}],["silent",{"0":{"22":1},"2":{"22":2}}],["singapore",{"2":{"19":1}}],["single",{"2":{"13":2,"19":1}}],["size",{"0":{"30":1,"33":1,"38":1},"1":{"31":1,"34":1,"39":1},"2":{"14":1,"15":4,"22":1,"29":4,"30":3,"31":2,"32":2,"33":3,"34":2,"38":3,"39":1,"57":1,"65":1}}],["successfully",{"2":{"14":1}}],["such",{"2":{"14":1}}],["supports",{"2":{"63":1}}],["support",{"2":{"13":1}}],["supported",{"2":{"0":1,"10":1,"11":1,"12":1}}],["svc3",{"2":{"13":1}}],["svc2",{"2":{"13":1}}],["svc1",{"2":{"13":1}}],["scans",{"2":{"15":1,"19":1}}],["scanning",{"2":{"13":1,"15":1,"40":1,"42":1}}],["scan",{"0":{"15":1},"2":{"13":1,"19":2,"31":1}}],["s",{"2":{"9":5,"14":1,"15":1,"17":1,"18":1,"40":1,"42":1,"60":1,"73":1}}],["sensitive",{"2":{"79":1,"80":1}}],["sends",{"2":{"45":2,"60":1,"83":2}}],["search",{"2":{"68":1}}],["searching",{"2":{"8":2}}],["set",{"2":{"51":2}}],["selected",{"2":{"45":1,"83":1}}],["seldomly",{"2":{"8":1}}],["securitygroupids",{"2":{"74":1}}],["security",{"0":{"73":1},"1":{"74":1,"75":1,"76":1},"2":{"15":1,"29":1,"42":1,"43":1,"73":3,"74":2,"75":1,"76":6,"90":1}}],["seconds",{"2":{"15":1,"35":1}}],["servers",{"2":{"21":1}}],["server",{"2":{"15":1}}],["serverless",{"2":{"14":3,"25":2,"26":2}}],["services",{"0":{"13":1},"2":{"13":2}}],["service",{"0":{"10":1},"1":{"11":1,"12":1,"13":1},"2":{"10":2}}],["several",{"2":{"13":1}}],["separated",{"2":{"13":1}}],["structured",{"2":{"63":2}}],["stick",{"2":{"51":1}}],["stored",{"2":{"48":1,"86":1}}],["storing",{"2":{"48":1,"86":1}}],["storage",{"0":{"30":1,"57":1},"1":{"31":1,"58":1},"2":{"15":2,"29":2,"30":3,"31":2,"57":2,"58":1}}],["startup",{"2":{"80":2,"81":1}}],["start",{"2":{"14":1,"34":1,"80":1}}],["started",{"0":{"14":1},"1":{"15":1,"16":1}}],["status",{"0":{"9":1,"77":1},"1":{"78":1,"79":1,"80":1,"81":1},"2":{"2":1,"9":2,"29":1,"77":2}}],["steps",{"0":{"16":1},"2":{"4":1,"37":1}}],["shorthand",{"2":{"19":1}}],["short",{"2":{"18":1,"20":1}}],["shot",{"2":{"13":1}}],["sh",{"2":{"1":1,"18":1,"19":1,"20":1,"28":1,"31":1}}],["sh$",{"2":{"0":1,"3":1,"5":1,"7":1,"9":1,"11":1,"12":1,"13":1,"14":2,"15":1,"21":1,"22":1,"25":2,"26":1,"27":1,"30":1,"32":1,"33":1,"35":1,"36":1,"38":1,"40":1,"42":1,"44":1,"49":1,"53":1,"54":1,"57":1,"59":1,"65":1,"67":1,"69":1,"71":1,"73":1,"77":1,"82":1,"87":1,"89":1}}]],"serializationVersion":2}';export{a as default};
