+++ docs/assets/chunks/VPLocalSearchBox.-BnIIPGv.js
@@ -1,4 +1,4 @@
-import{Y as pt,h as oe,y as Ve,ai as kt,aj as Nt,d as It,G as xe,ak as et,g as Fe,al as Dt,am as Ot,z as _t,an as Rt,j as Oe,P as he,W as Ee,ao as Mt,U as Lt,V as Pt,ap as zt,Z as Bt,v as Vt,aq as $t,o as ee,b as Wt,k as E,a2 as jt,m as U,ar as Kt,as as Jt,at as Ut,c as re,n as tt,e as Se,E as rt,F as at,a as ve,t as pe,au as Ht,p as Gt,q as qt,av as nt,aw as Qt,a8 as Yt,ae as Zt,_ as Xt}from"./framework.KK7BHIzh.js";import{u as er,c as tr,L as rr}from"./theme.jayLMn_d.js";const ar={root:()=>pt(()=>import("./@localSearchIndexroot.fsb5mFmL.js"),__vite__mapDeps([]))};/*!
+import{Y as pt,h as oe,y as Ve,ai as kt,aj as Nt,d as It,G as xe,ak as et,g as Fe,al as Dt,am as Ot,z as _t,an as Rt,j as Oe,P as he,W as Ee,ao as Mt,U as Lt,V as Pt,ap as zt,Z as Bt,v as Vt,aq as $t,o as ee,b as Wt,k as E,a2 as jt,m as U,ar as Kt,as as Jt,at as Ut,c as re,n as tt,e as Se,E as rt,F as at,a as ve,t as pe,au as Ht,p as Gt,q as qt,av as nt,aw as Qt,a8 as Yt,ae as Zt,_ as Xt}from"./framework.KK7BHIzh.js";import{u as er,c as tr,L as rr}from"./theme.mkGsDXPC.js";const ar={root:()=>pt(()=>import("./@localSearchIndexroot.fNJnBJgz.js"),__vite__mapDeps([]))};/*!
 * tabbable 6.2.0
 * @license MIT, https://github.com/focus-trap/tabbable/blob/master/LICENSE
 */var mt=["input:not([inert])","select:not([inert])","textarea:not([inert])","a[href]:not([inert])","button:not([inert])","[tabindex]:not(slot):not([inert])","audio[controls]:not([inert])","video[controls]:not([inert])",'[contenteditable]:not([contenteditable="false"]):not([inert])',"details>summary:first-of-type:not([inert])","details:not([inert])"],Te=mt.join(","),yt=typeof Element>"u",ue=yt?function(){}:Element.prototype.matches||Element.prototype.msMatchesSelector||Element.prototype.webkitMatchesSelector,ke=!yt&&Element.prototype.getRootNode?function(o){var e;return o==null||(e=o.getRootNode)===null||e===void 0?void 0:e.call(o)}:function(o){return o==null?void 0:o.ownerDocument},Ne=function o(e,t){var r;t===void 0&&(t=!0);var n=e==null||(r=e.getAttribute)===null||r===void 0?void 0:r.call(e,"inert"),a=n===""||n==="true",i=a||t&&e&&o(e.parentNode);return i},nr=function(e){var t,r=e==null||(t=e.getAttribute)===null||t===void 0?void 0:t.call(e,"contenteditable");return r===""||r==="true"},gt=function(e,t,r){if(Ne(e))return[];var n=Array.prototype.slice.apply(e.querySelectorAll(Te));return t&&ue.call(e,Te)&&n.unshift(e),n=n.filter(r),n},bt=function o(e,t,r){for(var n=[],a=Array.from(e);a.length;){var i=a.shift();if(!Ne(i,!1))if(i.tagName==="SLOT"){var s=i.assignedElements(),u=s.length?s:i.children,l=o(u,!0,r);r.flatten?n.push.apply(n,l):n.push({scopeParent:i,candidates:l})}else{var d=ue.call(i,Te);d&&r.filter(i)&&(t||!e.includes(i))&&n.push(i);var h=i.shadowRoot||typeof r.getShadowRoot=="function"&&r.getShadowRoot(i),v=!Ne(h,!1)&&(!r.shadowRootFilter||r.shadowRootFilter(i));if(h&&v){var y=o(h===!0?i.children:h.children,!0,r);r.flatten?n.push.apply(n,y):n.push({scopeParent:i,candidates:y})}else a.unshift.apply(a,i.children)}}return n},wt=function(e){return!isNaN(parseInt(e.getAttribute("tabindex"),10))},se=function(e){if(!e)throw new Error("No node provided");return e.tabIndex<0&&(/^(AUDIO|VIDEO|DETAILS)$/.test(e.tagName)||nr(e))&&!wt(e)?0:e.tabIndex},ir=function(e,t){var r=se(e);return r<0&&t&&!wt(e)?0:r},or=function(e,t){return e.tabIndex===t.tabIndex?e.documentOrder-t.documentOrder:e.tabIndex-t.tabIndex},xt=function(e){return e.tagName==="INPUT"},sr=function(e){return xt(e)&&e.type==="hidden"},ur=function(e){var t=e.tagName==="DETAILS"&&Array.prototype.slice.apply(e.children).some(function(r){return r.tagName==="SUMMARY"});return t},lr=function(e,t){for(var r=0;r<e.length;r++)if(e[r].checked&&e[r].form===t)return e[r]},cr=function(e){if(!e.name)return!0;var t=e.form||ke(e),r=function(s){return t.querySelectorAll('input[type="radio"][name="'+s+'"]')},n;if(typeof window<"u"&&typeof window.CSS<"u"&&typeof window.CSS.escape=="function")n=r(window.CSS.escape(e.name));else try{n=r(e.name)}catch(i){return console.error("Looks like you have a radio button with a name attribute containing invalid CSS selector characters and need the CSS.escape polyfill: %s",i.message),!1}var a=lr(n,e.form);return!a||a===e},fr=function(e){return xt(e)&&e.type==="radio"},dr=function(e){return fr(e)&&!cr(e)},hr=function(e){var t,r=e&&ke(e),n=(t=r)===null||t===void 0?void 0:t.host,a=!1;if(r&&r!==e){var i,s,u;for(a=!!((i=n)!==null&&i!==void 0&&(s=i.ownerDocument)!==null&&s!==void 0&&s.contains(n)||e!=null&&(u=e.ownerDocument)!==null&&u!==void 0&&u.contains(e));!a&&n;){var l,d,h;r=ke(n),n=(l=r)===null||l===void 0?void 0:l.host,a=!!((d=n)!==null&&d!==void 0&&(h=d.ownerDocument)!==null&&h!==void 0&&h.contains(n))}}return a},it=function(e){var t=e.getBoundingClientRect(),r=t.width,n=t.height;return r===0&&n===0},vr=function(e,t){var r=t.displayCheck,n=t.getShadowRoot;if(getComputedStyle(e).visibility==="hidden")return!0;var a=ue.call(e,"details>summary:first-of-type"),i=a?e.parentElement:e;if(ue.call(i,"details:not([open]) *"))return!0;if(!r||r==="full"||r==="legacy-full"){if(typeof n=="function"){for(var s=e;e;){var u=e.parentElement,l=ke(e);if(u&&!u.shadowRoot&&n(u)===!0)return it(e);e.assignedSlot?e=e.assignedSlot:!u&&l!==e.ownerDocument?e=l.host:e=u}e=s}if(hr(e))return!e.getClientRects().length;if(r!=="legacy-full")return!0}else if(r==="non-zero-area")return it(e);return!1},pr=function(e){if(/^(INPUT|BUTTON|SELECT|TEXTAREA)$/.test(e.tagName))for(var t=e.parentElement;t;){if(t.tagName==="FIELDSET"&&t.disabled){for(var r=0;r<t.children.length;r++){var n=t.children.item(r);if(n.tagName==="LEGEND")return ue.call(t,"fieldset[disabled] *")?!0:!n.contains(e)}return!0}t=t.parentElement}return!1},Ie=function(e,t){return!(t.disabled||Ne(t)||sr(t)||vr(t,e)||ur(t)||pr(t))},$e=function(e,t){return!(dr(t)||se(t)<0||!Ie(e,t))},mr=function(e){var t=parseInt(e.getAttribute("tabindex"),10);return!!(isNaN(t)||t>=0)},yr=function o(e){var t=[],r=[];return e.forEach(function(n,a){var i=!!n.scopeParent,s=i?n.scopeParent:n,u=ir(s,i),l=i?o(n.candidates):s;u===0?i?t.push.apply(t,l):t.push(s):r.push({documentOrder:a,tabIndex:u,item:n,isScope:i,content:l})}),r.sort(or).reduce(function(n,a){return a.isScope?n.push.apply(n,a.content):n.push(a.content),n},[]).concat(t)},gr=function(e,t){t=t||{};var r;return t.getShadowRoot?r=bt([e],t.includeContainer,{filter:$e.bind(null,t),flatten:!1,getShadowRoot:t.getShadowRoot,shadowRootFilter:mr}):r=gt(e,t.includeContainer,$e.bind(null,t)),yr(r)},br=function(e,t){t=t||{};var r;return t.getShadowRoot?r=bt([e],t.includeContainer,{filter:Ie.bind(null,t),flatten:!0,getShadowRoot:t.getShadowRoot}):r=gt(e,t.includeContainer,Ie.bind(null,t)),r},le=function(e,t){if(t=t||{},!e)throw new Error("No node provided");return ue.call(e,Te)===!1?!1:$e(t,e)},wr=mt.concat("iframe").join(","),_e=function(e,t){if(t=t||{},!e)throw new Error("No node provided");return ue.call(e,wr)===!1?!1:Ie(t,e)};/*!
