--- docs/assets/chunks/theme.jayLMn_d.js
@@ -1,7 +1,0 @@
-var De=Object.defineProperty;var Fe=(s,e,t)=>e in s?De(s,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):s[e]=t;var oe=(s,e,t)=>(Fe(s,typeof e!="symbol"?e+"":e,t),t);import{d as $,o as a,c as l,r as u,n as T,a as H,t as L,_ as m,b as k,w as h,T as ce,e as f,u as xe,i as Oe,l as Ue,f as ue,g as b,h as M,j as U,k as c,m as i,p as E,q as D,s as O,v as K,x as re,y as G,z as ee,A as de,B as Ve,C as Ge,D as j,F as C,E as B,G as ve,H as Y,I as _,J as x,K as Le,L as te,M as X,N as se,O as je,P as Re,Q as Se,R as qe,S as Ke,U as We,V as Ye,W as ye,X as Je,Y as Xe,Z as Me,$ as Ce,a0 as Ze,a1 as Qe,a2 as et,a3 as tt}from"./framework.KK7BHIzh.js";const st=$({__name:"VPBadge",props:{text:{},type:{default:"tip"}},setup(s){return(e,t)=>(a(),l("span",{class:T(["VPBadge",e.type])},[u(e.$slots,"default",{},()=>[H(L(e.text),1)],!0)],2))}}),nt=m(st,[["__scopeId","data-v-ea5b2908"]]),ot={key:0,class:"VPBackdrop"},at=$({__name:"VPBackdrop",props:{show:{type:Boolean}},setup(s){return(e,t)=>(a(),k(ce,{name:"fade"},{default:h(()=>[e.show?(a(),l("div",ot)):f("",!0)]),_:1}))}}),rt=m(at,[["__scopeId","data-v-54a304ca"]]),P=xe;function it(s,e){let t,n=!1;return()=>{t&&clearTimeout(t),n?t=setTimeout(s,e):(s(),(n=!0)&&setTimeout(()=>n=!1,e))}}function ie(s){return/^\//.test(s)?s:`/${s}`}function he(s){const{pathname:e,search:t,hash:n,protocol:o}=new URL(s,"http://a.com");if(Oe(s)||s.startsWith("#")||!o.startsWith("http")||/\.(?!html|md)\w+($|\?)/i.test(s)&&Ue(s))return s;const{site:r}=P(),d=e.endsWith("/")||e.endsWith(".html")?s:s.replace(/(?:(^\.+)\/)?.*$/,`$1${e.replace(/(\.md)?$/,r.value.cleanUrls?"":".html")}${t}${n}`);return ue(d)}function J({removeCurrent:s=!0,correspondingLink:e=!1}={}){const{site:t,localeIndex:n,page:o,theme:r}=P(),d=b(()=>{var v,g;return{label:(v=t.value.locales[n.value])==null?void 0:v.label,link:((g=t.value.locales[n.value])==null?void 0:g.link)||(n.value==="root"?"/":`/${n.value}/`)}});return{localeLinks:b(()=>Object.entries(t.value.locales).flatMap(([v,g])=>s&&d.value.label===g.label?[]:{text:g.label,link:lt(g.link||(v==="root"?"/":`/${v}/`),r.value.i18nRouting!==!1&&e,o.value.relativePath.slice(d.value.link.length-1),!t.value.cleanUrls)})),currentLang:d}}function lt(s,e,t,n){return e?s.replace(/\/$/,"")+ie(t.replace(/(^|\/)index\.md$/,"$1").replace(/\.md$/,n?".html":"")):s}const ct=s=>(E("data-v-b9c0c15a"),s=s(),D(),s),ut={class:"NotFound"},dt={class:"code"},vt={class:"title"},ht=ct(()=>c("div",{class:"divider"},null,-1)),pt={class:"quote"},_t={class:"action"},ft=["href","aria-label"],mt=$({__name:"NotFound",setup(s){const{site:e,theme:t}=P(),{localeLinks:n}=J({removeCurrent:!1}),o=M("/");return U(()=>{var d;const r=window.location.pathname.replace(e.value.base,"").replace(/(^.*?\/).*$/,"/$1");n.value.length&&(o.value=((d=n.value.find(({link:p})=>p.startsWith(r)))==null?void 0:d.link)||n.value[0].link)}),(r,d)=>{var p,v,g,w,y;return a(),l("div",ut,[c("p",dt,L(((p=i(t).notFound)==null?void 0:p.code)??"404"),1),c("h1",vt,L(((v=i(t).notFound)==null?void 0:v.title)??"PAGE NOT FOUND"),1),ht,c("blockquote",pt,L(((g=i(t).notFound)==null?void 0:g.quote)??"But if you don't change your direction, and if you keep looking, you may end up where you are heading."),1),c("div",_t,[c("a",{class:"link",href:i(ue)(o.value),"aria-label":((w=i(t).notFound)==null?void 0:w.linkLabel)??"go to home"},L(((y=i(t).notFound)==null?void 0:y.linkText)??"Take me home"),9,ft)])])}}}),gt=m(mt,[["__scopeId","data-v-b9c0c15a"]]);function Ie(s,e){if(Array.isArray(s))return Z(s);if(s==null)return[];e=ie(e);const t=Object.keys(s).sort((o,r)=>r.split("/").length-o.split("/").length).find(o=>e.startsWith(ie(o))),n=t?s[t]:[];return Array.isArray(n)?Z(n):Z(n.items,n.base)}function $t(s){const e=[];let t=0;for(const n in s){const o=s[n];if(o.items){t=e.push(o);continue}e[t]||e.push({items:[]}),e[t].items.push(o)}return e}function kt(s){const e=[];function t(n){for(const o of n)o.text&&o.link&&e.push({text:o.text,link:o.link,docFooterText:o.docFooterText}),o.items&&t(o.items)}return t(s),e}function le(s,e){return Array.isArray(e)?e.some(t=>le(s,t)):O(s,e.link)?!0:e.items?le(s,e.items):!1}function Z(s,e){return[...s].map(t=>{const n={...t},o=n.base||e;return o&&n.link&&(n.link=o+n.link),n.items&&(n.items=Z(n.items,o)),n})}function F(){const{frontmatter:s,page:e,theme:t}=P(),n=re("(min-width: 960px)"),o=M(!1),r=b(()=>{const A=t.value.sidebar,V=e.value.relativePath;return A?Ie(A,V):[]}),d=M(r.value);G(r,(A,V)=>{JSON.stringify(A)!==JSON.stringify(V)&&(d.value=r.value)});const p=b(()=>s.value.sidebar!==!1&&d.value.length>0&&s.value.layout!=="home"),v=b(()=>g?s.value.aside==null?t.value.aside==="left":s.value.aside==="left":!1),g=b(()=>s.value.layout==="home"?!1:s.value.aside!=null?!!s.value.aside:t.value.aside!==!1),w=b(()=>p.value&&n.value),y=b(()=>p.value?$t(d.value):[]);function I(){o.value=!0}function S(){o.value=!1}function N(){o.value?S():I()}return{isOpen:o,sidebar:d,sidebarGroups:y,hasSidebar:p,hasAside:g,leftAside:v,isSidebarEnabled:w,open:I,close:S,toggle:N}}function bt(s,e){let t;ee(()=>{t=s.value?document.activeElement:void 0}),U(()=>{window.addEventListener("keyup",n)}),de(()=>{window.removeEventListener("keyup",n)});function n(o){o.key==="Escape"&&s.value&&(e(),t==null||t.focus())}}const Te=M(K?location.hash:"");K&&window.addEventListener("hashchange",()=>{Te.value=location.hash});function yt(s){const{page:e}=P(),t=M(!1),n=b(()=>s.value.collapsed!=null),o=b(()=>!!s.value.link),r=M(!1),d=()=>{r.value=O(e.value.relativePath,s.value.link)};G([e,s,Te],d),U(d);const p=b(()=>r.value?!0:s.value.items?le(e.value.relativePath,s.value.items):!1),v=b(()=>!!(s.value.items&&s.value.items.length));ee(()=>{t.value=!!(n.value&&s.value.collapsed)}),Ve(()=>{(r.value||p.value)&&(t.value=!1)});function g(){n.value&&(t.value=!t.value)}return{collapsed:t,collapsible:n,isLink:o,isActiveLink:r,hasActiveLink:p,hasChildren:v,toggle:g}}function Pt(){const{hasSidebar:s}=F(),e=re("(min-width: 960px)"),t=re("(min-width: 1280px)");return{isAsideEnabled:b(()=>!t.value&&!e.value?!1:s.value?t.value:e.value)}}const wt=71;function pe(s){return typeof s.outline=="object"&&!Array.isArray(s.outline)&&s.outline.label||s.outlineTitle||"On this page"}function _e(s){const e=[...document.querySelectorAll(".VPDoc :where(h1,h2,h3,h4,h5,h6)")].filter(t=>t.id&&t.hasChildNodes()).map(t=>{const n=Number(t.tagName[1]);return{title:Vt(t),link:"#"+t.id,level:n}});return Lt(e,s)}function Vt(s){let e="";for(const t of s.childNodes)if(t.nodeType===1){if(t.classList.contains("VPBadge")||t.classList.contains("header-anchor"))continue;e+=t.textContent}else t.nodeType===3&&(e+=t.textContent);return e.trim()}function Lt(s,e){if(e===!1)return[];const t=(typeof e=="object"&&!Array.isArray(e)?e.level:e)||2,[n,o]=typeof t=="number"?[t,t]:t==="deep"?[2,6]:t;s=s.filter(d=>d.level>=n&&d.level<=o);const r=[];e:for(let d=0;d<s.length;d++){const p=s[d];if(d===0)r.push(p);else{for(let v=d-1;v>=0;v--){const g=s[v];if(g.level<p.level){(g.children||(g.children=[])).push(p);continue e}}r.push(p)}}return r}function St(s,e){const{isAsideEnabled:t}=Pt(),n=it(r,100);let o=null;U(()=>{requestAnimationFrame(r),window.addEventListener("scroll",n)}),Ge(()=>{d(location.hash)}),de(()=>{window.removeEventListener("scroll",n)});function r(){if(!t.value)return;const p=[].slice.call(s.value.querySelectorAll(".outline-link")),v=[].slice.call(document.querySelectorAll(".content .header-anchor")).filter(S=>p.some(N=>N.hash===S.hash&&S.offsetParent!==null)),g=window.scrollY,w=window.innerHeight,y=document.body.offsetHeight,I=Math.abs(g+w-y)<1;if(v.length&&I){d(v[v.length-1].hash);return}for(let S=0;S<v.length;S++){const N=v[S],A=v[S+1],[V,R]=Mt(S,N,A);if(V){d(R);return}}}function d(p){o&&o.classList.remove("active"),p==null?o=null:o=s.value.querySelector(`a[href="${decodeURIComponent(p)}"]`);const v=o;v?(v.classList.add("active"),e.value.style.top=v.offsetTop+33+"px",e.value.style.opacity="1"):(e.value.style.top="33px",e.value.style.opacity="0")}}function Pe(s){return s.parentElement.offsetTop-wt}function Mt(s,e,t){const n=window.scrollY;return s===0&&n===0?[!0,null]:n<Pe(e)?[!1,null]:!t||n<Pe(t)?[!0,e.hash]:[!1,null]}const Ct=["href","title"],It=$({__name:"VPDocOutlineItem",props:{headers:{},root:{type:Boolean}},setup(s){function e({target:t}){const n=t.href.split("#")[1],o=document.getElementById(decodeURIComponent(n));o==null||o.focus({preventScroll:!0})}return(t,n)=>{const o=j("VPDocOutlineItem",!0);return a(),l("ul",{class:T(t.root?"root":"nested")},[(a(!0),l(C,null,B(t.headers,({children:r,link:d,title:p})=>(a(),l("li",null,[c("a",{class:"outline-link",href:d,onClick:e,title:p},L(p),9,Ct),r!=null&&r.length?(a(),k(o,{key:0,headers:r},null,8,["headers"])):f("",!0)]))),256))],2)}}}),fe=m(It,[["__scopeId","data-v-463da30f"]]),Tt=s=>(E("data-v-3a6c4994"),s=s(),D(),s),At={class:"content"},Nt={class:"outline-title",role:"heading","aria-level":"2"},Bt={"aria-labelledby":"doc-outline-aria-label"},Ht=Tt(()=>c("span",{class:"visually-hidden",id:"doc-outline-aria-label"}," Table of Contents for current page ",-1)),zt=$({__name:"VPDocAsideOutline",setup(s){const{frontmatter:e,theme:t}=P(),n=ve([]);Y(()=>{n.value=_e(e.value.outline??t.value.outline)});const o=M(),r=M();return St(o,r),(d,p)=>(a(),l("div",{class:T(["VPDocAsideOutline",{"has-outline":n.value.length>0}]),ref_key:"container",ref:o,role:"navigation"},[c("div",At,[c("div",{class:"outline-marker",ref_key:"marker",ref:r},null,512),c("div",Nt,L(i(pe)(i(t))),1),c("nav",Bt,[Ht,_(fe,{headers:n.value,root:!0},null,8,["headers"])])])],2))}}),Et=m(zt,[["__scopeId","data-v-3a6c4994"]]),Dt={class:"VPDocAsideCarbonAds"},Ft=$({__name:"VPDocAsideCarbonAds",props:{carbonAds:{}},setup(s){const e=()=>null;return(t,n)=>(a(),l("div",Dt,[_(i(e),{"carbon-ads":t.carbonAds},null,8,["carbon-ads"])]))}}),xt=s=>(E("data-v-cb998dce"),s=s(),D(),s),Ot={class:"VPDocAside"},Ut=xt(()=>c("div",{class:"spacer"},null,-1)),Gt=$({__name:"VPDocAside",setup(s){const{theme:e}=P();return(t,n)=>(a(),l("div",Ot,[u(t.$slots,"aside-top",{},void 0,!0),u(t.$slots,"aside-outline-before",{},void 0,!0),_(Et),u(t.$slots,"aside-outline-after",{},void 0,!0),Ut,u(t.$slots,"aside-ads-before",{},void 0,!0),i(e).carbonAds?(a(),k(Ft,{key:0,"carbon-ads":i(e).carbonAds},null,8,["carbon-ads"])):f("",!0),u(t.$slots,"aside-ads-after",{},void 0,!0),u(t.$slots,"aside-bottom",{},void 0,!0)]))}}),jt=m(Gt,[["__scopeId","data-v-cb998dce"]]);function Rt(){const{theme:s,page:e}=P();return b(()=>{const{text:t="Edit this page",pattern:n=""}=s.value.editLink||{};let o;return typeof n=="function"?o=n(e.value):o=n.replace(/:path/g,e.value.filePath),{url:o,text:t}})}function qt(){const{page:s,theme:e,frontmatter:t}=P();return b(()=>{var v,g,w,y,I,S,N,A;const n=Ie(e.value.sidebar,s.value.relativePath),o=kt(n),r=o.findIndex(V=>O(s.value.relativePath,V.link)),d=((v=e.value.docFooter)==null?void 0:v.prev)===!1&&!t.value.prev||t.value.prev===!1,p=((g=e.value.docFooter)==null?void 0:g.next)===!1&&!t.value.next||t.value.next===!1;return{prev:d?void 0:{text:(typeof t.value.prev=="string"?t.value.prev:typeof t.value.prev=="object"?t.value.prev.text:void 0)??((w=o[r-1])==null?void 0:w.docFooterText)??((y=o[r-1])==null?void 0:y.text),link:(typeof t.value.prev=="object"?t.value.prev.link:void 0)??((I=o[r-1])==null?void 0:I.link)},next:p?void 0:{text:(typeof t.value.next=="string"?t.value.next:typeof t.value.next=="object"?t.value.next.text:void 0)??((S=o[r+1])==null?void 0:S.docFooterText)??((N=o[r+1])==null?void 0:N.text),link:(typeof t.value.next=="object"?t.value.next.link:void 0)??((A=o[r+1])==null?void 0:A.link)}}})}const Kt={},Wt={xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24"},Yt=c("path",{d:"M18,23H4c-1.7,0-3-1.3-3-3V6c0-1.7,1.3-3,3-3h7c0.6,0,1,0.4,1,1s-0.4,1-1,1H4C3.4,5,3,5.4,3,6v14c0,0.6,0.4,1,1,1h14c0.6,0,1-0.4,1-1v-7c0-0.6,0.4-1,1-1s1,0.4,1,1v7C21,21.7,19.7,23,18,23z"},null,-1),Jt=c("path",{d:"M8,17c-0.3,0-0.5-0.1-0.7-0.3C7,16.5,6.9,16.1,7,15.8l1-4c0-0.2,0.1-0.3,0.3-0.5l9.5-9.5c1.2-1.2,3.2-1.2,4.4,0c1.2,1.2,1.2,3.2,0,4.4l-9.5,9.5c-0.1,0.1-0.3,0.2-0.5,0.3l-4,1C8.2,17,8.1,17,8,17zM9.9,12.5l-0.5,2.1l2.1-0.5l9.3-9.3c0.4-0.4,0.4-1.1,0-1.6c-0.4-0.4-1.2-0.4-1.6,0l0,0L9.9,12.5z M18.5,2.5L18.5,2.5L18.5,2.5z"},null,-1),Xt=[Yt,Jt];function Zt(s,e){return a(),l("svg",Wt,Xt)}const Qt=m(Kt,[["render",Zt]]),z=$({__name:"VPLink",props:{tag:{},href:{},noIcon:{type:Boolean},target:{},rel:{}},setup(s){const e=s,t=b(()=>e.tag??(e.href?"a":"span")),n=b(()=>e.href&&Le.test(e.href));return(o,r)=>(a(),k(x(t.value),{class:T(["VPLink",{link:o.href,"vp-external-link-icon":n.value,"no-icon":o.noIcon}]),href:o.href?i(he)(o.href):void 0,target:o.target??(n.value?"_blank":void 0),rel:o.rel??(n.value?"noreferrer":void 0)},{default:h(()=>[u(o.$slots,"default")]),_:3},8,["class","href","target","rel"]))}}),es={class:"VPLastUpdated"},ts=["datetime"],ss=$({__name:"VPDocFooterLastUpdated",setup(s){const{theme:e,page:t,frontmatter:n,lang:o}=P(),r=b(()=>new Date(n.value.lastUpdated??t.value.lastUpdated)),d=b(()=>r.value.toISOString()),p=M("");return U(()=>{ee(()=>{var v,g,w;p.value=new Intl.DateTimeFormat((g=(v=e.value.lastUpdated)==null?void 0:v.formatOptions)!=null&&g.forceLocale?o.value:void 0,((w=e.value.lastUpdated)==null?void 0:w.formatOptions)??{dateStyle:"short",timeStyle:"short"}).format(r.value)})}),(v,g)=>{var w;return a(),l("p",es,[H(L(((w=i(e).lastUpdated)==null?void 0:w.text)||i(e).lastUpdatedText||"Last updated")+": ",1),c("time",{datetime:d.value},L(p.value),9,ts)])}}}),ns=m(ss,[["__scopeId","data-v-19a7ae4e"]]),os={key:0,class:"VPDocFooter"},as={key:0,class:"edit-info"},rs={key:0,class:"edit-link"},is={key:1,class:"last-updated"},ls={key:1,class:"prev-next"},cs={class:"pager"},us=["innerHTML"],ds=["innerHTML"],vs={class:"pager"},hs=["innerHTML"],ps=["innerHTML"],_s=$({__name:"VPDocFooter",setup(s){const{theme:e,page:t,frontmatter:n}=P(),o=Rt(),r=qt(),d=b(()=>e.value.editLink&&n.value.editLink!==!1),p=b(()=>t.value.lastUpdated&&n.value.lastUpdated!==!1),v=b(()=>d.value||p.value||r.value.prev||r.value.next);return(g,w)=>{var y,I,S,N;return v.value?(a(),l("footer",os,[u(g.$slots,"doc-footer-before",{},void 0,!0),d.value||p.value?(a(),l("div",as,[d.value?(a(),l("div",rs,[_(z,{class:"edit-link-button",href:i(o).url,"no-icon":!0},{default:h(()=>[_(Qt,{class:"edit-link-icon","aria-label":"edit icon"}),H(" "+L(i(o).text),1)]),_:1},8,["href"])])):f("",!0),p.value?(a(),l("div",is,[_(ns)])):f("",!0)])):f("",!0),(y=i(r).prev)!=null&&y.link||(I=i(r).next)!=null&&I.link?(a(),l("nav",ls,[c("div",cs,[(S=i(r).prev)!=null&&S.link?(a(),k(z,{key:0,class:"pager-link prev",href:i(r).prev.link},{default:h(()=>{var A;return[c("span",{class:"desc",innerHTML:((A=i(e).docFooter)==null?void 0:A.prev)||"Previous page"},null,8,us),c("span",{class:"title",innerHTML:i(r).prev.text},null,8,ds)]}),_:1},8,["href"])):f("",!0)]),c("div",vs,[(N=i(r).next)!=null&&N.link?(a(),k(z,{key:0,class:"pager-link next",href:i(r).next.link},{default:h(()=>{var A;return[c("span",{class:"desc",innerHTML:((A=i(e).docFooter)==null?void 0:A.next)||"Next page"},null,8,hs),c("span",{class:"title",innerHTML:i(r).next.text},null,8,ps)]}),_:1},8,["href"])):f("",!0)])])):f("",!0)])):f("",!0)}}}),fs=m(_s,[["__scopeId","data-v-b4b63abf"]]),ms={},gs={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},$s=c("path",{d:"M9,19c-0.3,0-0.5-0.1-0.7-0.3c-0.4-0.4-0.4-1,0-1.4l5.3-5.3L8.3,6.7c-0.4-0.4-0.4-1,0-1.4s1-0.4,1.4,0l6,6c0.4,0.4,0.4,1,0,1.4l-6,6C9.5,18.9,9.3,19,9,19z"},null,-1),ks=[$s];function bs(s,e){return a(),l("svg",gs,ks)}const me=m(ms,[["render",bs]]),ys={key:0,class:"VPDocOutlineDropdown"},Ps={key:0,class:"items"},ws=$({__name:"VPDocOutlineDropdown",setup(s){const{frontmatter:e,theme:t}=P(),n=M(!1);Y(()=>{n.value=!1});const o=ve([]);return Y(()=>{o.value=_e(e.value.outline??t.value.outline)}),(r,d)=>o.value.length>0?(a(),l("div",ys,[c("button",{onClick:d[0]||(d[0]=p=>n.value=!n.value),class:T({open:n.value})},[H(L(i(pe)(i(t)))+" ",1),_(me,{class:"icon"})],2),n.value?(a(),l("div",Ps,[_(fe,{headers:o.value},null,8,["headers"])])):f("",!0)])):f("",!0)}}),Vs=m(ws,[["__scopeId","data-v-95bb0785"]]),Ls=s=>(E("data-v-a3c25e27"),s=s(),D(),s),Ss={class:"container"},Ms=Ls(()=>c("div",{class:"aside-curtain"},null,-1)),Cs={class:"aside-container"},Is={class:"aside-content"},Ts={class:"content"},As={class:"content-container"},Ns={class:"main"},Bs=$({__name:"VPDoc",setup(s){const{theme:e}=P(),t=te(),{hasSidebar:n,hasAside:o,leftAside:r}=F(),d=b(()=>t.path.replace(/[./]+/g,"_").replace(/_html$/,""));return(p,v)=>{const g=j("Content");return a(),l("div",{class:T(["VPDoc",{"has-sidebar":i(n),"has-aside":i(o)}])},[u(p.$slots,"doc-top",{},void 0,!0),c("div",Ss,[i(o)?(a(),l("div",{key:0,class:T(["aside",{"left-aside":i(r)}])},[Ms,c("div",Cs,[c("div",Is,[_(jt,null,{"aside-top":h(()=>[u(p.$slots,"aside-top",{},void 0,!0)]),"aside-bottom":h(()=>[u(p.$slots,"aside-bottom",{},void 0,!0)]),"aside-outline-before":h(()=>[u(p.$slots,"aside-outline-before",{},void 0,!0)]),"aside-outline-after":h(()=>[u(p.$slots,"aside-outline-after",{},void 0,!0)]),"aside-ads-before":h(()=>[u(p.$slots,"aside-ads-before",{},void 0,!0)]),"aside-ads-after":h(()=>[u(p.$slots,"aside-ads-after",{},void 0,!0)]),_:3})])])],2)):f("",!0),c("div",Ts,[c("div",As,[u(p.$slots,"doc-before",{},void 0,!0),_(Vs),c("main",Ns,[_(g,{class:T(["vp-doc",[d.value,i(e).externalLinkIcon&&"external-link-icon-enabled"]])},null,8,["class"])]),_(fs,null,{"doc-footer-before":h(()=>[u(p.$slots,"doc-footer-before",{},void 0,!0)]),_:3}),u(p.$slots,"doc-after",{},void 0,!0)])])]),u(p.$slots,"doc-bottom",{},void 0,!0)],2)}}}),Hs=m(Bs,[["__scopeId","data-v-a3c25e27"]]),zs=$({__name:"VPButton",props:{tag:{},size:{default:"medium"},theme:{default:"brand"},text:{},href:{}},setup(s){const e=s,t=b(()=>e.href&&Le.test(e.href)),n=b(()=>e.tag||e.href?"a":"button");return(o,r)=>(a(),k(x(n.value),{class:T(["VPButton",[o.size,o.theme]]),href:o.href?i(he)(o.href):void 0,target:t.value?"_blank":void 0,rel:t.value?"noreferrer":void 0},{default:h(()=>[H(L(o.text),1)]),_:1},8,["class","href","target","rel"]))}}),Es=m(zs,[["__scopeId","data-v-1e76fe75"]]),Ds=["src","alt"],Fs=$({inheritAttrs:!1,__name:"VPImage",props:{image:{},alt:{}},setup(s){return(e,t)=>{const n=j("VPImage",!0);return e.image?(a(),l(C,{key:0},[typeof e.image=="string"||"src"in e.image?(a(),l("img",X({key:0,class:"VPImage"},typeof e.image=="string"?e.$attrs:{...e.image,...e.$attrs},{src:i(ue)(typeof e.image=="string"?e.image:e.image.src),alt:e.alt??(typeof e.image=="string"?"":e.image.alt||"")}),null,16,Ds)):(a(),l(C,{key:1},[_(n,X({class:"dark",image:e.image.dark,alt:e.image.alt},e.$attrs),null,16,["image","alt"]),_(n,X({class:"light",image:e.image.light,alt:e.image.alt},e.$attrs),null,16,["image","alt"])],64))],64)):f("",!0)}}}),Q=m(Fs,[["__scopeId","data-v-ab19afbb"]]),xs=s=>(E("data-v-5a3e9999"),s=s(),D(),s),Os={class:"container"},Us={class:"main"},Gs={key:0,class:"name"},js=["innerHTML"],Rs=["innerHTML"],qs=["innerHTML"],Ks={key:0,class:"actions"},Ws={key:0,class:"image"},Ys={class:"image-container"},Js=xs(()=>c("div",{class:"image-bg"},null,-1)),Xs=$({__name:"VPHero",props:{name:{},text:{},tagline:{},image:{},actions:{}},setup(s){const e=se("hero-image-slot-exists");return(t,n)=>(a(),l("div",{class:T(["VPHero",{"has-image":t.image||i(e)}])},[c("div",Os,[c("div",Us,[u(t.$slots,"home-hero-info",{},()=>[t.name?(a(),l("h1",Gs,[c("span",{innerHTML:t.name,class:"clip"},null,8,js)])):f("",!0),t.text?(a(),l("p",{key:1,innerHTML:t.text,class:"text"},null,8,Rs)):f("",!0),t.tagline?(a(),l("p",{key:2,innerHTML:t.tagline,class:"tagline"},null,8,qs)):f("",!0)],!0),t.actions?(a(),l("div",Ks,[(a(!0),l(C,null,B(t.actions,o=>(a(),l("div",{key:o.link,class:"action"},[_(Es,{tag:"a",size:"medium",theme:o.theme,text:o.text,href:o.link},null,8,["theme","text","href"])]))),128))])):f("",!0)]),t.image||i(e)?(a(),l("div",Ws,[c("div",Ys,[Js,u(t.$slots,"home-hero-image",{},()=>[t.image?(a(),k(Q,{key:0,class:"image-src",image:t.image},null,8,["image"])):f("",!0)],!0)])])):f("",!0)])],2))}}),Zs=m(Xs,[["__scopeId","data-v-5a3e9999"]]),Qs=$({__name:"VPHomeHero",setup(s){const{frontmatter:e}=P();return(t,n)=>i(e).hero?(a(),k(Zs,{key:0,class:"VPHomeHero",name:i(e).hero.name,text:i(e).hero.text,tagline:i(e).hero.tagline,image:i(e).hero.image,actions:i(e).hero.actions},{"home-hero-info":h(()=>[u(t.$slots,"home-hero-info")]),"home-hero-image":h(()=>[u(t.$slots,"home-hero-image")]),_:3},8,["name","text","tagline","image","actions"])):f("",!0)}}),en={},tn={xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24"},sn=c("path",{d:"M19.9,12.4c0.1-0.2,0.1-0.5,0-0.8c-0.1-0.1-0.1-0.2-0.2-0.3l-7-7c-0.4-0.4-1-0.4-1.4,0s-0.4,1,0,1.4l5.3,5.3H5c-0.6,0-1,0.4-1,1s0.4,1,1,1h11.6l-5.3,5.3c-0.4,0.4-0.4,1,0,1.4c0.2,0.2,0.5,0.3,0.7,0.3s0.5-0.1,0.7-0.3l7-7C19.8,12.6,19.9,12.5,19.9,12.4z"},null,-1),nn=[sn];function on(s,e){return a(),l("svg",tn,nn)}const an=m(en,[["render",on]]),rn={class:"box"},ln={key:0,class:"icon"},cn=["innerHTML"],un=["innerHTML"],dn=["innerHTML"],vn={key:4,class:"link-text"},hn={class:"link-text-value"},pn=$({__name:"VPFeature",props:{icon:{},title:{},details:{},link:{},linkText:{},rel:{},target:{}},setup(s){return(e,t)=>(a(),k(z,{class:"VPFeature",href:e.link,rel:e.rel,target:e.target,"no-icon":!0,tag:e.link?"a":"div"},{default:h(()=>[c("article",rn,[typeof e.icon=="object"&&e.icon.wrap?(a(),l("div",ln,[_(Q,{image:e.icon,alt:e.icon.alt,height:e.icon.height||48,width:e.icon.width||48},null,8,["image","alt","height","width"])])):typeof e.icon=="object"?(a(),k(Q,{key:1,image:e.icon,alt:e.icon.alt,height:e.icon.height||48,width:e.icon.width||48},null,8,["image","alt","height","width"])):e.icon?(a(),l("div",{key:2,class:"icon",innerHTML:e.icon},null,8,cn)):f("",!0),c("h2",{class:"title",innerHTML:e.title},null,8,un),e.details?(a(),l("p",{key:3,class:"details",innerHTML:e.details},null,8,dn)):f("",!0),e.linkText?(a(),l("div",vn,[c("p",hn,[H(L(e.linkText)+" ",1),_(an,{class:"link-text-icon"})])])):f("",!0)])]),_:1},8,["href","rel","target","tag"]))}}),_n=m(pn,[["__scopeId","data-v-ee984185"]]),fn={key:0,class:"VPFeatures"},mn={class:"container"},gn={class:"items"},$n=$({__name:"VPFeatures",props:{features:{}},setup(s){const e=s,t=b(()=>{const n=e.features.length;if(n){if(n===2)return"grid-2";if(n===3)return"grid-3";if(n%3===0)return"grid-6";if(n>3)return"grid-4"}else return});return(n,o)=>n.features?(a(),l("div",fn,[c("div",mn,[c("div",gn,[(a(!0),l(C,null,B(n.features,r=>(a(),l("div",{key:r.title,class:T(["item",[t.value]])},[_(_n,{icon:r.icon,title:r.title,details:r.details,link:r.link,"link-text":r.linkText,rel:r.rel,target:r.target},null,8,["icon","title","details","link","link-text","rel","target"])],2))),128))])])])):f("",!0)}}),kn=m($n,[["__scopeId","data-v-b1eea84a"]]),bn=$({__name:"VPHomeFeatures",setup(s){const{frontmatter:e}=P();return(t,n)=>i(e).features?(a(),k(kn,{key:0,class:"VPHomeFeatures",features:i(e).features},null,8,["features"])):f("",!0)}}),yn={class:"VPHome"},Pn=$({__name:"VPHome",setup(s){return(e,t)=>{const n=j("Content");return a(),l("div",yn,[u(e.$slots,"home-hero-before",{},void 0,!0),_(Qs,null,{"home-hero-info":h(()=>[u(e.$slots,"home-hero-info",{},void 0,!0)]),"home-hero-image":h(()=>[u(e.$slots,"home-hero-image",{},void 0,!0)]),_:3}),u(e.$slots,"home-hero-after",{},void 0,!0),u(e.$slots,"home-features-before",{},void 0,!0),_(bn),u(e.$slots,"home-features-after",{},void 0,!0),_(n)])}}}),wn=m(Pn,[["__scopeId","data-v-20eabd3a"]]),Vn={},Ln={class:"VPPage"};function Sn(s,e){const t=j("Content");return a(),l("div",Ln,[u(s.$slots,"page-top"),_(t),u(s.$slots,"page-bottom")])}const Mn=m(Vn,[["render",Sn]]),Cn=$({__name:"VPContent",setup(s){const{page:e,frontmatter:t}=P(),{hasSidebar:n}=F();return(o,r)=>(a(),l("div",{class:T(["VPContent",{"has-sidebar":i(n),"is-home":i(t).layout==="home"}]),id:"VPContent"},[i(e).isNotFound?u(o.$slots,"not-found",{key:0},()=>[_(gt)],!0):i(t).layout==="page"?(a(),k(Mn,{key:1},{"page-top":h(()=>[u(o.$slots,"page-top",{},void 0,!0)]),"page-bottom":h(()=>[u(o.$slots,"page-bottom",{},void 0,!0)]),_:3})):i(t).layout==="home"?(a(),k(wn,{key:2},{"home-hero-before":h(()=>[u(o.$slots,"home-hero-before",{},void 0,!0)]),"home-hero-info":h(()=>[u(o.$slots,"home-hero-info",{},void 0,!0)]),"home-hero-image":h(()=>[u(o.$slots,"home-hero-image",{},void 0,!0)]),"home-hero-after":h(()=>[u(o.$slots,"home-hero-after",{},void 0,!0)]),"home-features-before":h(()=>[u(o.$slots,"home-features-before",{},void 0,!0)]),"home-features-after":h(()=>[u(o.$slots,"home-features-after",{},void 0,!0)]),_:3})):i(t).layout&&i(t).layout!=="doc"?(a(),k(x(i(t).layout),{key:3})):(a(),k(Hs,{key:4},{"doc-top":h(()=>[u(o.$slots,"doc-top",{},void 0,!0)]),"doc-bottom":h(()=>[u(o.$slots,"doc-bottom",{},void 0,!0)]),"doc-footer-before":h(()=>[u(o.$slots,"doc-footer-before",{},void 0,!0)]),"doc-before":h(()=>[u(o.$slots,"doc-before",{},void 0,!0)]),"doc-after":h(()=>[u(o.$slots,"doc-after",{},void 0,!0)]),"aside-top":h(()=>[u(o.$slots,"aside-top",{},void 0,!0)]),"aside-outline-before":h(()=>[u(o.$slots,"aside-outline-before",{},void 0,!0)]),"aside-outline-after":h(()=>[u(o.$slots,"aside-outline-after",{},void 0,!0)]),"aside-ads-before":h(()=>[u(o.$slots,"aside-ads-before",{},void 0,!0)]),"aside-ads-after":h(()=>[u(o.$slots,"aside-ads-after",{},void 0,!0)]),"aside-bottom":h(()=>[u(o.$slots,"aside-bottom",{},void 0,!0)]),_:3}))],2))}}),In=m(Cn,[["__scopeId","data-v-3cf691b6"]]),Tn={class:"container"},An=["innerHTML"],Nn=["innerHTML"],Bn=$({__name:"VPFooter",setup(s){const{theme:e,frontmatter:t}=P(),{hasSidebar:n}=F();return(o,r)=>i(e).footer&&i(t).footer!==!1?(a(),l("footer",{key:0,class:T(["VPFooter",{"has-sidebar":i(n)}])},[c("div",Tn,[i(e).footer.message?(a(),l("p",{key:0,class:"message",innerHTML:i(e).footer.message},null,8,An)):f("",!0),i(e).footer.copyright?(a(),l("p",{key:1,class:"copyright",innerHTML:i(e).footer.copyright},null,8,Nn)):f("",!0)])],2)):f("",!0)}}),Hn=m(Bn,[["__scopeId","data-v-566314d4"]]),zn={class:"header"},En={class:"outline"},Dn=$({__name:"VPLocalNavOutlineDropdown",props:{headers:{},navHeight:{}},setup(s){const e=s,{theme:t}=P(),n=M(!1),o=M(0),r=M();Y(()=>{n.value=!1});function d(){n.value=!n.value,o.value=window.innerHeight+Math.min(window.scrollY-e.navHeight,0)}function p(g){g.target.classList.contains("outline-link")&&(r.value&&(r.value.style.transition="none"),Re(()=>{n.value=!1}))}function v(){n.value=!1,window.scrollTo({top:0,left:0,behavior:"smooth"})}return(g,w)=>(a(),l("div",{class:"VPLocalNavOutlineDropdown",style:je({"--vp-vh":o.value+"px"})},[g.headers.length>0?(a(),l("button",{key:0,onClick:d,class:T({open:n.value})},[H(L(i(pe)(i(t)))+" ",1),_(me,{class:"icon"})],2)):(a(),l("button",{key:1,onClick:v},L(i(t).returnToTopLabel||"Return to top"),1)),_(ce,{name:"flyout"},{default:h(()=>[n.value?(a(),l("div",{key:0,ref_key:"items",ref:r,class:"items",onClick:p},[c("div",zn,[c("a",{class:"top-link",href:"#",onClick:v},L(i(t).returnToTopLabel||"Return to top"),1)]),c("div",En,[_(fe,{headers:g.headers},null,8,["headers"])])],512)):f("",!0)]),_:1})],4))}}),Fn=m(Dn,[["__scopeId","data-v-24251f6f"]]),xn={},On={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},Un=c("path",{d:"M17,11H3c-0.6,0-1-0.4-1-1s0.4-1,1-1h14c0.6,0,1,0.4,1,1S17.6,11,17,11z"},null,-1),Gn=c("path",{d:"M21,7H3C2.4,7,2,6.6,2,6s0.4-1,1-1h18c0.6,0,1,0.4,1,1S21.6,7,21,7z"},null,-1),jn=c("path",{d:"M21,15H3c-0.6,0-1-0.4-1-1s0.4-1,1-1h18c0.6,0,1,0.4,1,1S21.6,15,21,15z"},null,-1),Rn=c("path",{d:"M17,19H3c-0.6,0-1-0.4-1-1s0.4-1,1-1h14c0.6,0,1,0.4,1,1S17.6,19,17,19z"},null,-1),qn=[Un,Gn,jn,Rn];function Kn(s,e){return a(),l("svg",On,qn)}const Wn=m(xn,[["render",Kn]]),Yn=["aria-expanded"],Jn={class:"menu-text"},Xn=$({__name:"VPLocalNav",props:{open:{type:Boolean}},emits:["open-menu"],setup(s){const{theme:e,frontmatter:t}=P(),{hasSidebar:n}=F(),{y:o}=Se(),r=ve([]),d=M(0);U(()=>{d.value=parseInt(getComputedStyle(document.documentElement).getPropertyValue("--vp-nav-height"))}),Y(()=>{r.value=_e(t.value.outline??e.value.outline)});const p=b(()=>r.value.length===0&&!n.value),v=b(()=>({VPLocalNav:!0,fixed:p.value,"reached-top":o.value>=d.value}));return(g,w)=>i(t).layout!=="home"&&(!p.value||i(o)>=d.value)?(a(),l("div",{key:0,class:T(v.value)},[i(n)?(a(),l("button",{key:0,class:"menu","aria-expanded":g.open,"aria-controls":"VPSidebarNav",onClick:w[0]||(w[0]=y=>g.$emit("open-menu"))},[_(Wn,{class:"menu-icon"}),c("span",Jn,L(i(e).sidebarMenuLabel||"Menu"),1)],8,Yn)):f("",!0),_(Fn,{headers:r.value,navHeight:d.value},null,8,["headers","navHeight"])],2)):f("",!0)}}),Zn=m(Xn,[["__scopeId","data-v-f8a0b38a"]]);function Qn(){const s=M(!1);function e(){s.value=!0,window.addEventListener("resize",o)}function t(){s.value=!1,window.removeEventListener("resize",o)}function n(){s.value?t():e()}function o(){window.outerWidth>=768&&t()}const r=te();return G(()=>r.path,t),{isScreenOpen:s,openScreen:e,closeScreen:t,toggleScreen:n}}const eo={},to={class:"VPSwitch",type:"button",role:"switch"},so={class:"check"},no={key:0,class:"icon"};function oo(s,e){return a(),l("button",to,[c("span",so,[s.$slots.default?(a(),l("span",no,[u(s.$slots,"default",{},void 0,!0)])):f("",!0)])])}const ao=m(eo,[["render",oo],["__scopeId","data-v-1c29e291"]]),ro={},io={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},lo=c("path",{d:"M12.1,22c-0.3,0-0.6,0-0.9,0c-5.5-0.5-9.5-5.4-9-10.9c0.4-4.8,4.2-8.6,9-9c0.4,0,0.8,0.2,1,0.5c0.2,0.3,0.2,0.8-0.1,1.1c-2,2.7-1.4,6.4,1.3,8.4c2.1,1.6,5,1.6,7.1,0c0.3-0.2,0.7-0.3,1.1-0.1c0.3,0.2,0.5,0.6,0.5,1c-0.2,2.7-1.5,5.1-3.6,6.8C16.6,21.2,14.4,22,12.1,22zM9.3,4.4c-2.9,1-5,3.6-5.2,6.8c-0.4,4.4,2.8,8.3,7.2,8.7c2.1,0.2,4.2-0.4,5.8-1.8c1.1-0.9,1.9-2.1,2.4-3.4c-2.5,0.9-5.3,0.5-7.5-1.1C9.2,11.4,8.1,7.7,9.3,4.4z"},null,-1),co=[lo];function uo(s,e){return a(),l("svg",io,co)}const vo=m(ro,[["render",uo]]),ho={},po={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},_o=qe('<path d="M12,18c-3.3,0-6-2.7-6-6s2.7-6,6-6s6,2.7,6,6S15.3,18,12,18zM12,8c-2.2,0-4,1.8-4,4c0,2.2,1.8,4,4,4c2.2,0,4-1.8,4-4C16,9.8,14.2,8,12,8z"></path><path d="M12,4c-0.6,0-1-0.4-1-1V1c0-0.6,0.4-1,1-1s1,0.4,1,1v2C13,3.6,12.6,4,12,4z"></path><path d="M12,24c-0.6,0-1-0.4-1-1v-2c0-0.6,0.4-1,1-1s1,0.4,1,1v2C13,23.6,12.6,24,12,24z"></path><path d="M5.6,6.6c-0.3,0-0.5-0.1-0.7-0.3L3.5,4.9c-0.4-0.4-0.4-1,0-1.4s1-0.4,1.4,0l1.4,1.4c0.4,0.4,0.4,1,0,1.4C6.2,6.5,5.9,6.6,5.6,6.6z"></path><path d="M19.8,20.8c-0.3,0-0.5-0.1-0.7-0.3l-1.4-1.4c-0.4-0.4-0.4-1,0-1.4s1-0.4,1.4,0l1.4,1.4c0.4,0.4,0.4,1,0,1.4C20.3,20.7,20,20.8,19.8,20.8z"></path><path d="M3,13H1c-0.6,0-1-0.4-1-1s0.4-1,1-1h2c0.6,0,1,0.4,1,1S3.6,13,3,13z"></path><path d="M23,13h-2c-0.6,0-1-0.4-1-1s0.4-1,1-1h2c0.6,0,1,0.4,1,1S23.6,13,23,13z"></path><path d="M4.2,20.8c-0.3,0-0.5-0.1-0.7-0.3c-0.4-0.4-0.4-1,0-1.4l1.4-1.4c0.4-0.4,1-0.4,1.4,0s0.4,1,0,1.4l-1.4,1.4C4.7,20.7,4.5,20.8,4.2,20.8z"></path><path d="M18.4,6.6c-0.3,0-0.5-0.1-0.7-0.3c-0.4-0.4-0.4-1,0-1.4l1.4-1.4c0.4-0.4,1-0.4,1.4,0s0.4,1,0,1.4l-1.4,1.4C18.9,6.5,18.6,6.6,18.4,6.6z"></path>',9),fo=[_o];function mo(s,e){return a(),l("svg",po,fo)}const go=m(ho,[["render",mo]]),$o=$({__name:"VPSwitchAppearance",setup(s){const{isDark:e}=P(),t=se("toggle-appearance",()=>{e.value=!e.value}),n=b(()=>e.value?"Switch to light theme":"Switch to dark theme");return(o,r)=>(a(),k(ao,{title:n.value,class:"VPSwitchAppearance","aria-checked":i(e),onClick:i(t)},{default:h(()=>[_(go,{class:"sun"}),_(vo,{class:"moon"})]),_:1},8,["title","aria-checked","onClick"]))}}),ge=m($o,[["__scopeId","data-v-70af5d02"]]),ko={key:0,class:"VPNavBarAppearance"},bo=$({__name:"VPNavBarAppearance",setup(s){const{site:e}=P();return(t,n)=>i(e).appearance&&i(e).appearance!=="force-dark"?(a(),l("div",ko,[_(ge)])):f("",!0)}}),yo=m(bo,[["__scopeId","data-v-283b26e9"]]),$e=M();let Ae=!1,ae=0;function Po(s){const e=M(!1);if(K){!Ae&&wo(),ae++;const t=G($e,n=>{var o,r,d;n===s.el.value||(o=s.el.value)!=null&&o.contains(n)?(e.value=!0,(r=s.onFocus)==null||r.call(s)):(e.value=!1,(d=s.onBlur)==null||d.call(s))});de(()=>{t(),ae--,ae||Vo()})}return Ke(e)}function wo(){document.addEventListener("focusin",Ne),Ae=!0,$e.value=document.activeElement}function Vo(){document.removeEventListener("focusin",Ne)}function Ne(){$e.value=document.activeElement}const Lo={},So={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},Mo=c("path",{d:"M12,16c-0.3,0-0.5-0.1-0.7-0.3l-6-6c-0.4-0.4-0.4-1,0-1.4s1-0.4,1.4,0l5.3,5.3l5.3-5.3c0.4-0.4,1-0.4,1.4,0s0.4,1,0,1.4l-6,6C12.5,15.9,12.3,16,12,16z"},null,-1),Co=[Mo];function Io(s,e){return a(),l("svg",So,Co)}const Be=m(Lo,[["render",Io]]),To={},Ao={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},No=c("circle",{cx:"12",cy:"12",r:"2"},null,-1),Bo=c("circle",{cx:"19",cy:"12",r:"2"},null,-1),Ho=c("circle",{cx:"5",cy:"12",r:"2"},null,-1),zo=[No,Bo,Ho];function Eo(s,e){return a(),l("svg",Ao,zo)}const Do=m(To,[["render",Eo]]),Fo={class:"VPMenuLink"},xo=$({__name:"VPMenuLink",props:{item:{}},setup(s){const{page:e}=P();return(t,n)=>(a(),l("div",Fo,[_(z,{class:T({active:i(O)(i(e).relativePath,t.item.activeMatch||t.item.link,!!t.item.activeMatch)}),href:t.item.link,target:t.item.target,rel:t.item.rel},{default:h(()=>[H(L(t.item.text),1)]),_:1},8,["class","href","target","rel"])]))}}),ne=m(xo,[["__scopeId","data-v-f51f088d"]]),Oo={class:"VPMenuGroup"},Uo={key:0,class:"title"},Go=$({__name:"VPMenuGroup",props:{text:{},items:{}},setup(s){return(e,t)=>(a(),l("div",Oo,[e.text?(a(),l("p",Uo,L(e.text),1)):f("",!0),(a(!0),l(C,null,B(e.items,n=>(a(),l(C,null,["link"in n?(a(),k(ne,{key:0,item:n},null,8,["item"])):f("",!0)],64))),256))]))}}),jo=m(Go,[["__scopeId","data-v-a6b0397c"]]),Ro={class:"VPMenu"},qo={key:0,class:"items"},Ko=$({__name:"VPMenu",props:{items:{}},setup(s){return(e,t)=>(a(),l("div",Ro,[e.items?(a(),l("div",qo,[(a(!0),l(C,null,B(e.items,n=>(a(),l(C,{key:n.text},["link"in n?(a(),k(ne,{key:0,item:n},null,8,["item"])):(a(),k(jo,{key:1,text:n.text,items:n.items},null,8,["text","items"]))],64))),128))])):f("",!0),u(e.$slots,"default",{},void 0,!0)]))}}),Wo=m(Ko,[["__scopeId","data-v-e42ed9b3"]]),Yo=["aria-expanded","aria-label"],Jo={key:0,class:"text"},Xo=["innerHTML"],Zo={class:"menu"},Qo=$({__name:"VPFlyout",props:{icon:{},button:{},label:{},items:{}},setup(s){const e=M(!1),t=M();Po({el:t,onBlur:n});function n(){e.value=!1}return(o,r)=>(a(),l("div",{class:"VPFlyout",ref_key:"el",ref:t,onMouseenter:r[1]||(r[1]=d=>e.value=!0),onMouseleave:r[2]||(r[2]=d=>e.value=!1)},[c("button",{type:"button",class:"button","aria-haspopup":"true","aria-expanded":e.value,"aria-label":o.label,onClick:r[0]||(r[0]=d=>e.value=!e.value)},[o.button||o.icon?(a(),l("span",Jo,[o.icon?(a(),k(x(o.icon),{key:0,class:"option-icon"})):f("",!0),o.button?(a(),l("span",{key:1,innerHTML:o.button},null,8,Xo)):f("",!0),_(Be,{class:"text-icon"})])):(a(),k(Do,{key:1,class:"icon"}))],8,Yo),c("div",Zo,[_(Wo,{items:o.items},{default:h(()=>[u(o.$slots,"default",{},void 0,!0)]),_:3},8,["items"])])],544))}}),ke=m(Qo,[["__scopeId","data-v-aa8de344"]]),ea={discord:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>Discord</title><path d="M20.317 4.3698a19.7913 19.7913 0 00-4.8851-1.5152.0741.0741 0 00-.0785.0371c-.211.3753-.4447.8648-.6083 1.2495-1.8447-.2762-3.68-.2762-5.4868 0-.1636-.3933-.4058-.8742-.6177-1.2495a.077.077 0 00-.0785-.037 19.7363 19.7363 0 00-4.8852 1.515.0699.0699 0 00-.0321.0277C.5334 9.0458-.319 13.5799.0992 18.0578a.0824.0824 0 00.0312.0561c2.0528 1.5076 4.0413 2.4228 5.9929 3.0294a.0777.0777 0 00.0842-.0276c.4616-.6304.8731-1.2952 1.226-1.9942a.076.076 0 00-.0416-.1057c-.6528-.2476-1.2743-.5495-1.8722-.8923a.077.077 0 01-.0076-.1277c.1258-.0943.2517-.1923.3718-.2914a.0743.0743 0 01.0776-.0105c3.9278 1.7933 8.18 1.7933 12.0614 0a.0739.0739 0 01.0785.0095c.1202.099.246.1981.3728.2924a.077.077 0 01-.0066.1276 12.2986 12.2986 0 01-1.873.8914.0766.0766 0 00-.0407.1067c.3604.698.7719 1.3628 1.225 1.9932a.076.076 0 00.0842.0286c1.961-.6067 3.9495-1.5219 6.0023-3.0294a.077.077 0 00.0313-.0552c.5004-5.177-.8382-9.6739-3.5485-13.6604a.061.061 0 00-.0312-.0286zM8.02 15.3312c-1.1825 0-2.1569-1.0857-2.1569-2.419 0-1.3332.9555-2.4189 2.157-2.4189 1.2108 0 2.1757 1.0952 2.1568 2.419 0 1.3332-.9555 2.4189-2.1569 2.4189zm7.9748 0c-1.1825 0-2.1569-1.0857-2.1569-2.419 0-1.3332.9554-2.4189 2.1569-2.4189 1.2108 0 2.1757 1.0952 2.1568 2.419 0 1.3332-.946 2.4189-2.1568 2.4189Z"/></svg>',facebook:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>Facebook</title><path d="M24 12.073c0-6.627-5.373-12-12-12s-12 5.373-12 12c0 5.99 4.388 10.954 10.125 11.854v-8.385H7.078v-3.47h3.047V9.43c0-3.007 1.792-4.669 4.533-4.669 1.312 0 2.686.235 2.686.235v2.953H15.83c-1.491 0-1.956.925-1.956 1.874v2.25h3.328l-.532 3.47h-2.796v8.385C19.612 23.027 24 18.062 24 12.073z"/></svg>',github:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>GitHub</title><path d="M12 .297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C4.422 18.07 3.633 17.7 3.633 17.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 22.092 24 17.592 24 12.297c0-6.627-5.373-12-12-12"/></svg>',instagram:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>Instagram</title><path d="M12 0C8.74 0 8.333.015 7.053.072 5.775.132 4.905.333 4.14.63c-.789.306-1.459.717-2.126 1.384S.935 3.35.63 4.14C.333 4.905.131 5.775.072 7.053.012 8.333 0 8.74 0 12s.015 3.667.072 4.947c.06 1.277.261 2.148.558 2.913.306.788.717 1.459 1.384 2.126.667.666 1.336 1.079 2.126 1.384.766.296 1.636.499 2.913.558C8.333 23.988 8.74 24 12 24s3.667-.015 4.947-.072c1.277-.06 2.148-.262 2.913-.558.788-.306 1.459-.718 2.126-1.384.666-.667 1.079-1.335 1.384-2.126.296-.765.499-1.636.558-2.913.06-1.28.072-1.687.072-4.947s-.015-3.667-.072-4.947c-.06-1.277-.262-2.149-.558-2.913-.306-.789-.718-1.459-1.384-2.126C21.319 1.347 20.651.935 19.86.63c-.765-.297-1.636-.499-2.913-.558C15.667.012 15.26 0 12 0zm0 2.16c3.203 0 3.585.016 4.85.071 1.17.055 1.805.249 2.227.415.562.217.96.477 1.382.896.419.42.679.819.896 1.381.164.422.36 1.057.413 2.227.057 1.266.07 1.646.07 4.85s-.015 3.585-.074 4.85c-.061 1.17-.256 1.805-.421 2.227-.224.562-.479.96-.899 1.382-.419.419-.824.679-1.38.896-.42.164-1.065.36-2.235.413-1.274.057-1.649.07-4.859.07-3.211 0-3.586-.015-4.859-.074-1.171-.061-1.816-.256-2.236-.421-.569-.224-.96-.479-1.379-.899-.421-.419-.69-.824-.9-1.38-.165-.42-.359-1.065-.42-2.235-.045-1.26-.061-1.649-.061-4.844 0-3.196.016-3.586.061-4.861.061-1.17.255-1.814.42-2.234.21-.57.479-.96.9-1.381.419-.419.81-.689 1.379-.898.42-.166 1.051-.361 2.221-.421 1.275-.045 1.65-.06 4.859-.06l.045.03zm0 3.678c-3.405 0-6.162 2.76-6.162 6.162 0 3.405 2.76 6.162 6.162 6.162 3.405 0 6.162-2.76 6.162-6.162 0-3.405-2.76-6.162-6.162-6.162zM12 16c-2.21 0-4-1.79-4-4s1.79-4 4-4 4 1.79 4 4-1.79 4-4 4zm7.846-10.405c0 .795-.646 1.44-1.44 1.44-.795 0-1.44-.646-1.44-1.44 0-.794.646-1.439 1.44-1.439.793-.001 1.44.645 1.44 1.439z"/></svg>',linkedin:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>LinkedIn</title><path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z"/></svg>',mastodon:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>Mastodon</title><path d="M23.268 5.313c-.35-2.578-2.617-4.61-5.304-5.004C17.51.242 15.792 0 11.813 0h-.03c-3.98 0-4.835.242-5.288.309C3.882.692 1.496 2.518.917 5.127.64 6.412.61 7.837.661 9.143c.074 1.874.088 3.745.26 5.611.118 1.24.325 2.47.62 3.68.55 2.237 2.777 4.098 4.96 4.857 2.336.792 4.849.923 7.256.38.265-.061.527-.132.786-.213.585-.184 1.27-.39 1.774-.753a.057.057 0 0 0 .023-.043v-1.809a.052.052 0 0 0-.02-.041.053.053 0 0 0-.046-.01 20.282 20.282 0 0 1-4.709.545c-2.73 0-3.463-1.284-3.674-1.818a5.593 5.593 0 0 1-.319-1.433.053.053 0 0 1 .066-.054c1.517.363 3.072.546 4.632.546.376 0 .75 0 1.125-.01 1.57-.044 3.224-.124 4.768-.422.038-.008.077-.015.11-.024 2.435-.464 4.753-1.92 4.989-5.604.008-.145.03-1.52.03-1.67.002-.512.167-3.63-.024-5.545zm-3.748 9.195h-2.561V8.29c0-1.309-.55-1.976-1.67-1.976-1.23 0-1.846.79-1.846 2.35v3.403h-2.546V8.663c0-1.56-.617-2.35-1.848-2.35-1.112 0-1.668.668-1.67 1.977v6.218H4.822V8.102c0-1.31.337-2.35 1.011-3.12.696-.77 1.608-1.164 2.74-1.164 1.311 0 2.302.5 2.962 1.498l.638 1.06.638-1.06c.66-.999 1.65-1.498 2.96-1.498 1.13 0 2.043.395 2.74 1.164.675.77 1.012 1.81 1.012 3.12z"/></svg>',slack:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>Slack</title><path d="M5.042 15.165a2.528 2.528 0 0 1-2.52 2.523A2.528 2.528 0 0 1 0 15.165a2.527 2.527 0 0 1 2.522-2.52h2.52v2.52zM6.313 15.165a2.527 2.527 0 0 1 2.521-2.52 2.527 2.527 0 0 1 2.521 2.52v6.313A2.528 2.528 0 0 1 8.834 24a2.528 2.528 0 0 1-2.521-2.522v-6.313zM8.834 5.042a2.528 2.528 0 0 1-2.521-2.52A2.528 2.528 0 0 1 8.834 0a2.528 2.528 0 0 1 2.521 2.522v2.52H8.834zM8.834 6.313a2.528 2.528 0 0 1 2.521 2.521 2.528 2.528 0 0 1-2.521 2.521H2.522A2.528 2.528 0 0 1 0 8.834a2.528 2.528 0 0 1 2.522-2.521h6.312zM18.956 8.834a2.528 2.528 0 0 1 2.522-2.521A2.528 2.528 0 0 1 24 8.834a2.528 2.528 0 0 1-2.522 2.521h-2.522V8.834zM17.688 8.834a2.528 2.528 0 0 1-2.523 2.521 2.527 2.527 0 0 1-2.52-2.521V2.522A2.527 2.527 0 0 1 15.165 0a2.528 2.528 0 0 1 2.523 2.522v6.312zM15.165 18.956a2.528 2.528 0 0 1 2.523 2.522A2.528 2.528 0 0 1 15.165 24a2.527 2.527 0 0 1-2.52-2.522v-2.522h2.52zM15.165 17.688a2.527 2.527 0 0 1-2.52-2.523 2.526 2.526 0 0 1 2.52-2.52h6.313A2.527 2.527 0 0 1 24 15.165a2.528 2.528 0 0 1-2.522 2.523h-6.313z"/></svg>',twitter:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>Twitter</title><path d="M21.543 7.104c.015.211.015.423.015.636 0 6.507-4.954 14.01-14.01 14.01v-.003A13.94 13.94 0 0 1 0 19.539a9.88 9.88 0 0 0 7.287-2.041 4.93 4.93 0 0 1-4.6-3.42 4.916 4.916 0 0 0 2.223-.084A4.926 4.926 0 0 1 .96 9.167v-.062a4.887 4.887 0 0 0 2.235.616A4.928 4.928 0 0 1 1.67 3.148 13.98 13.98 0 0 0 11.82 8.292a4.929 4.929 0 0 1 8.39-4.49 9.868 9.868 0 0 0 3.128-1.196 4.941 4.941 0 0 1-2.165 2.724A9.828 9.828 0 0 0 24 4.555a10.019 10.019 0 0 1-2.457 2.549z"/></svg>',x:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>X</title><path d="M18.901 1.153h3.68l-8.04 9.19L24 22.846h-7.406l-5.8-7.584-6.638 7.584H.474l8.6-9.83L0 1.154h7.594l5.243 6.932ZM17.61 20.644h2.039L6.486 3.24H4.298Z"/></svg>',youtube:'<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>YouTube</title><path d="M23.498 6.186a3.016 3.016 0 0 0-2.122-2.136C19.505 3.545 12 3.545 12 3.545s-7.505 0-9.377.505A3.017 3.017 0 0 0 .502 6.186C0 8.07 0 12 0 12s0 3.93.502 5.814a3.016 3.016 0 0 0 2.122 2.136c1.871.505 9.376.505 9.376.505s7.505 0 9.377-.505a3.015 3.015 0 0 0 2.122-2.136C24 15.93 24 12 24 12s0-3.93-.502-5.814zM9.545 15.568V8.432L15.818 12l-6.273 3.568z"/></svg>'},ta=["href","aria-label","innerHTML"],sa=$({__name:"VPSocialLink",props:{icon:{},link:{},ariaLabel:{}},setup(s){const e=s,t=b(()=>typeof e.icon=="object"?e.icon.svg:ea[e.icon]);return(n,o)=>(a(),l("a",{class:"VPSocialLink no-icon",href:n.link,"aria-label":n.ariaLabel??(typeof n.icon=="string"?n.icon:""),target:"_blank",rel:"noopener",innerHTML:t.value},null,8,ta))}}),na=m(sa,[["__scopeId","data-v-16cf740a"]]),oa={class:"VPSocialLinks"},aa=$({__name:"VPSocialLinks",props:{links:{}},setup(s){return(e,t)=>(a(),l("div",oa,[(a(!0),l(C,null,B(e.links,({link:n,icon:o,ariaLabel:r})=>(a(),k(na,{key:n,icon:o,link:n,ariaLabel:r},null,8,["icon","link","ariaLabel"]))),128))]))}}),be=m(aa,[["__scopeId","data-v-e71e869c"]]),ra={key:0,class:"group translations"},ia={class:"trans-title"},la={key:1,class:"group"},ca={class:"item appearance"},ua={class:"label"},da={class:"appearance-action"},va={key:2,class:"group"},ha={class:"item social-links"},pa=$({__name:"VPNavBarExtra",setup(s){const{site:e,theme:t}=P(),{localeLinks:n,currentLang:o}=J({correspondingLink:!0}),r=b(()=>n.value.length&&o.value.label||e.value.appearance||t.value.socialLinks);return(d,p)=>r.value?(a(),k(ke,{key:0,class:"VPNavBarExtra",label:"extra navigation"},{default:h(()=>[i(n).length&&i(o).label?(a(),l("div",ra,[c("p",ia,L(i(o).label),1),(a(!0),l(C,null,B(i(n),v=>(a(),k(ne,{key:v.link,item:v},null,8,["item"]))),128))])):f("",!0),i(e).appearance&&i(e).appearance!=="force-dark"?(a(),l("div",la,[c("div",ca,[c("p",ua,L(i(t).darkModeSwitchLabel||"Appearance"),1),c("div",da,[_(ge)])])])):f("",!0),i(t).socialLinks?(a(),l("div",va,[c("div",ha,[_(be,{class:"social-links-list",links:i(t).socialLinks},null,8,["links"])])])):f("",!0)]),_:1})):f("",!0)}}),_a=m(pa,[["__scopeId","data-v-8e87c032"]]),fa=s=>(E("data-v-6bee1efd"),s=s(),D(),s),ma=["aria-expanded"],ga=fa(()=>c("span",{class:"container"},[c("span",{class:"top"}),c("span",{class:"middle"}),c("span",{class:"bottom"})],-1)),$a=[ga],ka=$({__name:"VPNavBarHamburger",props:{active:{type:Boolean}},emits:["click"],setup(s){return(e,t)=>(a(),l("button",{type:"button",class:T(["VPNavBarHamburger",{active:e.active}]),"aria-label":"mobile navigation","aria-expanded":e.active,"aria-controls":"VPNavScreen",onClick:t[0]||(t[0]=n=>e.$emit("click"))},$a,10,ma))}}),ba=m(ka,[["__scopeId","data-v-6bee1efd"]]),ya=["innerHTML"],Pa=$({__name:"VPNavBarMenuLink",props:{item:{}},setup(s){const{page:e}=P();return(t,n)=>(a(),k(z,{class:T({VPNavBarMenuLink:!0,active:i(O)(i(e).relativePath,t.item.activeMatch||t.item.link,!!t.item.activeMatch)}),href:t.item.link,target:t.item.target,rel:t.item.rel,tabindex:"0"},{default:h(()=>[c("span",{innerHTML:t.item.text},null,8,ya)]),_:1},8,["class","href","target","rel"]))}}),wa=m(Pa,[["__scopeId","data-v-cb318fec"]]),Va=$({__name:"VPNavBarMenuGroup",props:{item:{}},setup(s){const e=s,{page:t}=P(),n=r=>"link"in r?O(t.value.relativePath,r.link,!!e.item.activeMatch):r.items.some(n),o=b(()=>n(e.item));return(r,d)=>(a(),k(ke,{class:T({VPNavBarMenuGroup:!0,active:i(O)(i(t).relativePath,r.item.activeMatch,!!r.item.activeMatch)||o.value}),button:r.item.text,items:r.item.items},null,8,["class","button","items"]))}}),La=s=>(E("data-v-f732b5d0"),s=s(),D(),s),Sa={key:0,"aria-labelledby":"main-nav-aria-label",class:"VPNavBarMenu"},Ma=La(()=>c("span",{id:"main-nav-aria-label",class:"visually-hidden"},"Main Navigation",-1)),Ca=$({__name:"VPNavBarMenu",setup(s){const{theme:e}=P();return(t,n)=>i(e).nav?(a(),l("nav",Sa,[Ma,(a(!0),l(C,null,B(i(e).nav,o=>(a(),l(C,{key:o.text},["link"in o?(a(),k(wa,{key:0,item:o},null,8,["item"])):(a(),k(Va,{key:1,item:o},null,8,["item"]))],64))),128))])):f("",!0)}}),Ia=m(Ca,[["__scopeId","data-v-f732b5d0"]]);function Ta(s,e){const{localeIndex:t}=P();function n(o){var S,N;const r=o.split("."),d=s&&typeof s=="object",p=d&&((N=(S=s.locales)==null?void 0:S[t.value])==null?void 0:N.translations)||null,v=d&&s.translations||null;let g=p,w=v,y=e;const I=r.pop();for(const A of r){let V=null;const R=y==null?void 0:y[A];R&&(V=y=R);const W=w==null?void 0:w[A];W&&(V=w=W);const q=g==null?void 0:g[A];q&&(V=g=q),R||(y=V),W||(w=V),q||(g=V)}return(g==null?void 0:g[I])??(w==null?void 0:w[I])??(y==null?void 0:y[I])??""}return n}const Aa=["aria-label"],Na={class:"DocSearch-Button-Container"},Ba=c("svg",{class:"DocSearch-Search-Icon",width:"20",height:"20",viewBox:"0 0 20 20","aria-label":"search icon"},[c("path",{d:"M14.386 14.386l4.0877 4.0877-4.0877-4.0877c-2.9418 2.9419-7.7115 2.9419-10.6533 0-2.9419-2.9418-2.9419-7.7115 0-10.6533 2.9418-2.9419 7.7115-2.9419 10.6533 0 2.9419 2.9418 2.9419 7.7115 0 10.6533z",stroke:"currentColor",fill:"none","fill-rule":"evenodd","stroke-linecap":"round","stroke-linejoin":"round"})],-1),Ha={class:"DocSearch-Button-Placeholder"},za=c("span",{class:"DocSearch-Button-Keys"},[c("kbd",{class:"DocSearch-Button-Key"}),c("kbd",{class:"DocSearch-Button-Key"},"K")],-1),we=$({__name:"VPNavBarSearchButton",setup(s){const{theme:e}=P(),t={button:{buttonText:"Search",buttonAriaLabel:"Search"}},n=We(Ta)(Ye(()=>{var o;return(o=e.value.search)==null?void 0:o.options}),t);return(o,r)=>(a(),l("button",{type:"button",class:"DocSearch DocSearch-Button","aria-label":i(n)("button.buttonAriaLabel")},[c("span",Na,[Ba,c("span",Ha,L(i(n)("button.buttonText")),1)]),za],8,Aa))}}),Ea={class:"VPNavBarSearch"},Da={id:"local-search"},Fa={key:1,id:"docsearch"},xa=$({__name:"VPNavBarSearch",setup(s){const e=Je(()=>Xe(()=>import("./VPLocalSearchBox.5p5Pr0YC.js"),__vite__mapDeps([0,1]))),t=()=>null,{theme:n}=P(),o=M(!1),r=M(!1);U(()=>{});function d(){o.value||(o.value=!0,setTimeout(p,16))}function p(){const y=new Event("keydown");y.key="k",y.metaKey=!0,window.dispatchEvent(y),setTimeout(()=>{document.querySelector(".DocSearch-Modal")||p()},16)}function v(y){const I=y.target,S=I.tagName;return I.isContentEditable||S==="INPUT"||S==="SELECT"||S==="TEXTAREA"}const g=M(!1);ye("k",y=>{(y.ctrlKey||y.metaKey)&&(y.preventDefault(),g.value=!0)}),ye("/",y=>{v(y)||(y.preventDefault(),g.value=!0)});const w="local";return(y,I)=>{var S;return a(),l("div",Ea,[i(w)==="local"?(a(),l(C,{key:0},[g.value?(a(),k(i(e),{key:0,onClose:I[0]||(I[0]=N=>g.value=!1)})):f("",!0),c("div",Da,[_(we,{onClick:I[1]||(I[1]=N=>g.value=!0)})])],64)):i(w)==="algolia"?(a(),l(C,{key:1},[o.value?(a(),k(i(t),{key:0,algolia:((S=i(n).search)==null?void 0:S.options)??i(n).algolia,onVnodeBeforeMount:I[2]||(I[2]=N=>r.value=!0)},null,8,["algolia"])):f("",!0),r.value?f("",!0):(a(),l("div",Fa,[_(we,{onClick:d})]))],64)):f("",!0)])}}}),Oa=$({__name:"VPNavBarSocialLinks",setup(s){const{theme:e}=P();return(t,n)=>i(e).socialLinks?(a(),k(be,{key:0,class:"VPNavBarSocialLinks",links:i(e).socialLinks},null,8,["links"])):f("",!0)}}),Ua=m(Oa,[["__scopeId","data-v-ef6192dc"]]),Ga=["href"],ja=$({__name:"VPNavBarTitle",setup(s){const{site:e,theme:t}=P(),{hasSidebar:n}=F(),{currentLang:o}=J();return(r,d)=>(a(),l("div",{class:T(["VPNavBarTitle",{"has-sidebar":i(n)}])},[c("a",{class:"title",href:i(t).logoLink??i(he)(i(o).link)},[u(r.$slots,"nav-bar-title-before",{},void 0,!0),i(t).logo?(a(),k(Q,{key:0,class:"logo",image:i(t).logo},null,8,["image"])):f("",!0),i(t).siteTitle?(a(),l(C,{key:1},[H(L(i(t).siteTitle),1)],64)):i(t).siteTitle===void 0?(a(),l(C,{key:2},[H(L(i(e).title),1)],64)):f("",!0),u(r.$slots,"nav-bar-title-after",{},void 0,!0)],8,Ga)],2))}}),Ra=m(ja,[["__scopeId","data-v-2973dbb4"]]),qa={},Ka={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},Wa=c("path",{d:"M0 0h24v24H0z",fill:"none"},null,-1),Ya=c("path",{d:" M12.87 15.07l-2.54-2.51.03-.03c1.74-1.94 2.98-4.17 3.71-6.53H17V4h-7V2H8v2H1v1.99h11.17C11.5 7.92 10.44 9.75 9 11.35 8.07 10.32 7.3 9.19 6.69 8h-2c.73 1.63 1.73 3.17 2.98 4.56l-5.09 5.02L4 19l5-5 3.11 3.11.76-2.04zM18.5 10h-2L12 22h2l1.12-3h4.75L21 22h2l-4.5-12zm-2.62 7l1.62-4.33L19.12 17h-3.24z ",class:"css-c4d79v"},null,-1),Ja=[Wa,Ya];function Xa(s,e){return a(),l("svg",Ka,Ja)}const He=m(qa,[["render",Xa]]),Za={class:"items"},Qa={class:"title"},er=$({__name:"VPNavBarTranslations",setup(s){const{theme:e}=P(),{localeLinks:t,currentLang:n}=J({correspondingLink:!0});return(o,r)=>i(t).length&&i(n).label?(a(),k(ke,{key:0,class:"VPNavBarTranslations",icon:He,label:i(e).langMenuLabel||"Change language"},{default:h(()=>[c("div",Za,[c("p",Qa,L(i(n).label),1),(a(!0),l(C,null,B(i(t),d=>(a(),k(ne,{key:d.link,item:d},null,8,["item"]))),128))])]),_:1},8,["label"])):f("",!0)}}),tr=m(er,[["__scopeId","data-v-ff4524ae"]]),sr=s=>(E("data-v-5befd255"),s=s(),D(),s),nr={class:"container"},or={class:"title"},ar={class:"content"},rr=sr(()=>c("div",{class:"curtain"},null,-1)),ir={class:"content-body"},lr=$({__name:"VPNavBar",props:{isScreenOpen:{type:Boolean}},emits:["toggle-screen"],setup(s){const{y:e}=Se(),{hasSidebar:t}=F(),{frontmatter:n}=P(),o=M({});return Ve(()=>{o.value={"has-sidebar":t.value,top:n.value.layout==="home"&&e.value===0}}),(r,d)=>(a(),l("div",{class:T(["VPNavBar",o.value])},[c("div",nr,[c("div",or,[_(Ra,null,{"nav-bar-title-before":h(()=>[u(r.$slots,"nav-bar-title-before",{},void 0,!0)]),"nav-bar-title-after":h(()=>[u(r.$slots,"nav-bar-title-after",{},void 0,!0)]),_:3})]),c("div",ar,[rr,c("div",ir,[u(r.$slots,"nav-bar-content-before",{},void 0,!0),_(xa,{class:"search"}),_(Ia,{class:"menu"}),_(tr,{class:"translations"}),_(yo,{class:"appearance"}),_(Ua,{class:"social-links"}),_(_a,{class:"extra"}),u(r.$slots,"nav-bar-content-after",{},void 0,!0),_(ba,{class:"hamburger",active:r.isScreenOpen,onClick:d[0]||(d[0]=p=>r.$emit("toggle-screen"))},null,8,["active"])])])])],2))}}),cr=m(lr,[["__scopeId","data-v-5befd255"]]),ur={key:0,class:"VPNavScreenAppearance"},dr={class:"text"},vr=$({__name:"VPNavScreenAppearance",setup(s){const{site:e,theme:t}=P();return(n,o)=>i(e).appearance&&i(e).appearance!=="force-dark"?(a(),l("div",ur,[c("p",dr,L(i(t).darkModeSwitchLabel||"Appearance"),1),_(ge)])):f("",!0)}}),hr=m(vr,[["__scopeId","data-v-338d9b48"]]),pr=$({__name:"VPNavScreenMenuLink",props:{item:{}},setup(s){const e=se("close-screen");return(t,n)=>(a(),k(z,{class:"VPNavScreenMenuLink",href:t.item.link,target:t.item.target,rel:t.item.rel,onClick:i(e)},{default:h(()=>[H(L(t.item.text),1)]),_:1},8,["href","target","rel","onClick"]))}}),_r=m(pr,[["__scopeId","data-v-fe523e3d"]]),fr={},mr={xmlns:"http://www.w3.org/2000/svg","aria-hidden":"true",focusable:"false",viewBox:"0 0 24 24"},gr=c("path",{d:"M18.9,10.9h-6v-6c0-0.6-0.4-1-1-1s-1,0.4-1,1v6h-6c-0.6,0-1,0.4-1,1s0.4,1,1,1h6v6c0,0.6,0.4,1,1,1s1-0.4,1-1v-6h6c0.6,0,1-0.4,1-1S19.5,10.9,18.9,10.9z"},null,-1),$r=[gr];function kr(s,e){return a(),l("svg",mr,$r)}const br=m(fr,[["render",kr]]),yr=$({__name:"VPNavScreenMenuGroupLink",props:{item:{}},setup(s){const e=se("close-screen");return(t,n)=>(a(),k(z,{class:"VPNavScreenMenuGroupLink",href:t.item.link,target:t.item.target,rel:t.item.rel,onClick:i(e)},{default:h(()=>[H(L(t.item.text),1)]),_:1},8,["href","target","rel","onClick"]))}}),ze=m(yr,[["__scopeId","data-v-aea78dd1"]]),Pr={class:"VPNavScreenMenuGroupSection"},wr={key:0,class:"title"},Vr=$({__name:"VPNavScreenMenuGroupSection",props:{text:{},items:{}},setup(s){return(e,t)=>(a(),l("div",Pr,[e.text?(a(),l("p",wr,L(e.text),1)):f("",!0),(a(!0),l(C,null,B(e.items,n=>(a(),k(ze,{key:n.text,item:n},null,8,["item"]))),128))]))}}),Lr=m(Vr,[["__scopeId","data-v-f60dbfa7"]]),Sr=["aria-controls","aria-expanded"],Mr=["innerHTML"],Cr=["id"],Ir={key:1,class:"group"},Tr=$({__name:"VPNavScreenMenuGroup",props:{text:{},items:{}},setup(s){const e=s,t=M(!1),n=b(()=>`NavScreenGroup-${e.text.replace(" ","-").toLowerCase()}`);function o(){t.value=!t.value}return(r,d)=>(a(),l("div",{class:T(["VPNavScreenMenuGroup",{open:t.value}])},[c("button",{class:"button","aria-controls":n.value,"aria-expanded":t.value,onClick:o},[c("span",{class:"button-text",innerHTML:r.text},null,8,Mr),_(br,{class:"button-icon"})],8,Sr),c("div",{id:n.value,class:"items"},[(a(!0),l(C,null,B(r.items,p=>(a(),l(C,{key:p.text},["link"in p?(a(),l("div",{key:p.text,class:"item"},[_(ze,{item:p},null,8,["item"])])):(a(),l("div",Ir,[_(Lr,{text:p.text,items:p.items},null,8,["text","items"])]))],64))),128))],8,Cr)],2))}}),Ar=m(Tr,[["__scopeId","data-v-32e4a89c"]]),Nr={key:0,class:"VPNavScreenMenu"},Br=$({__name:"VPNavScreenMenu",setup(s){const{theme:e}=P();return(t,n)=>i(e).nav?(a(),l("nav",Nr,[(a(!0),l(C,null,B(i(e).nav,o=>(a(),l(C,{key:o.text},["link"in o?(a(),k(_r,{key:0,item:o},null,8,["item"])):(a(),k(Ar,{key:1,text:o.text||"",items:o.items},null,8,["text","items"]))],64))),128))])):f("",!0)}}),Hr=$({__name:"VPNavScreenSocialLinks",setup(s){const{theme:e}=P();return(t,n)=>i(e).socialLinks?(a(),k(be,{key:0,class:"VPNavScreenSocialLinks",links:i(e).socialLinks},null,8,["links"])):f("",!0)}}),zr={class:"list"},Er=$({__name:"VPNavScreenTranslations",setup(s){const{localeLinks:e,currentLang:t}=J({correspondingLink:!0}),n=M(!1);function o(){n.value=!n.value}return(r,d)=>i(e).length&&i(t).label?(a(),l("div",{key:0,class:T(["VPNavScreenTranslations",{open:n.value}])},[c("button",{class:"title",onClick:o},[_(He,{class:"icon lang"}),H(" "+L(i(t).label)+" ",1),_(Be,{class:"icon chevron"})]),c("ul",zr,[(a(!0),l(C,null,B(i(e),p=>(a(),l("li",{key:p.link,class:"item"},[_(z,{class:"link",href:p.link},{default:h(()=>[H(L(p.text),1)]),_:2},1032,["href"])]))),128))])],2)):f("",!0)}}),Dr=m(Er,[["__scopeId","data-v-41505286"]]),Fr={class:"container"},xr=$({__name:"VPNavScreen",props:{open:{type:Boolean}},setup(s){const e=M(null),t=Me(K?document.body:null);return(n,o)=>(a(),k(ce,{name:"fade",onEnter:o[0]||(o[0]=r=>t.value=!0),onAfterLeave:o[1]||(o[1]=r=>t.value=!1)},{default:h(()=>[n.open?(a(),l("div",{key:0,class:"VPNavScreen",ref_key:"screen",ref:e,id:"VPNavScreen"},[c("div",Fr,[u(n.$slots,"nav-screen-content-before",{},void 0,!0),_(Br,{class:"menu"}),_(Dr,{class:"translations"}),_(hr,{class:"appearance"}),_(Hr,{class:"social-links"}),u(n.$slots,"nav-screen-content-after",{},void 0,!0)])],512)):f("",!0)]),_:3}))}}),Or=m(xr,[["__scopeId","data-v-57cce842"]]),Ur={key:0,class:"VPNav"},Gr=$({__name:"VPNav",setup(s){const{isScreenOpen:e,closeScreen:t,toggleScreen:n}=Qn(),{frontmatter:o}=P(),r=b(()=>o.value.navbar!==!1);return Ce("close-screen",t),ee(()=>{K&&document.documentElement.classList.toggle("hide-nav",!r.value)}),(d,p)=>r.value?(a(),l("header",Ur,[_(cr,{"is-screen-open":i(e),onToggleScreen:i(n)},{"nav-bar-title-before":h(()=>[u(d.$slots,"nav-bar-title-before",{},void 0,!0)]),"nav-bar-title-after":h(()=>[u(d.$slots,"nav-bar-title-after",{},void 0,!0)]),"nav-bar-content-before":h(()=>[u(d.$slots,"nav-bar-content-before",{},void 0,!0)]),"nav-bar-content-after":h(()=>[u(d.$slots,"nav-bar-content-after",{},void 0,!0)]),_:3},8,["is-screen-open","onToggleScreen"]),_(Or,{open:i(e)},{"nav-screen-content-before":h(()=>[u(d.$slots,"nav-screen-content-before",{},void 0,!0)]),"nav-screen-content-after":h(()=>[u(d.$slots,"nav-screen-content-after",{},void 0,!0)]),_:3},8,["open"])])):f("",!0)}}),jr=m(Gr,[["__scopeId","data-v-7ad780c2"]]),Rr=s=>(E("data-v-bd01e0d5"),s=s(),D(),s),qr=["role","tabindex"],Kr=Rr(()=>c("div",{class:"indicator"},null,-1)),Wr={key:1,class:"items"},Yr=$({__name:"VPSidebarItem",props:{item:{},depth:{}},setup(s){const e=s,{collapsed:t,collapsible:n,isLink:o,isActiveLink:r,hasActiveLink:d,hasChildren:p,toggle:v}=yt(b(()=>e.item)),g=b(()=>p.value?"section":"div"),w=b(()=>o.value?"a":"div"),y=b(()=>p.value?e.depth+2===7?"p":`h${e.depth+2}`:"p"),I=b(()=>o.value?void 0:"button"),S=b(()=>[[`level-${e.depth}`],{collapsible:n.value},{collapsed:t.value},{"is-link":o.value},{"is-active":r.value},{"has-active":d.value}]);function N(V){"key"in V&&V.key!=="Enter"||!e.item.link&&v()}function A(){e.item.link&&v()}return(V,R)=>{const W=j("VPSidebarItem",!0);return a(),k(x(g.value),{class:T(["VPSidebarItem",S.value])},{default:h(()=>[V.item.text?(a(),l("div",X({key:0,class:"item",role:I.value},Qe(V.item.items?{click:N,keydown:N}:{},!0),{tabindex:V.item.items&&0}),[Kr,V.item.link?(a(),k(z,{key:0,tag:w.value,class:"link",href:V.item.link,rel:V.item.rel,target:V.item.target},{default:h(()=>[(a(),k(x(y.value),{class:"text",innerHTML:V.item.text},null,8,["innerHTML"]))]),_:1},8,["tag","href","rel","target"])):(a(),k(x(y.value),{key:1,class:"text",innerHTML:V.item.text},null,8,["innerHTML"])),V.item.collapsed!=null?(a(),l("div",{key:2,class:"caret",role:"button","aria-label":"toggle section",onClick:A,onKeydown:Ze(A,["enter"]),tabindex:"0"},[_(me,{class:"caret-icon"})],32)):f("",!0)],16,qr)):f("",!0),V.item.items&&V.item.items.length?(a(),l("div",Wr,[V.depth<5?(a(!0),l(C,{key:0},B(V.item.items,q=>(a(),k(W,{key:q.text,item:q,depth:V.depth+1},null,8,["item","depth"]))),128)):f("",!0)])):f("",!0)]),_:1},8,["class"])}}}),Jr=m(Yr,[["__scopeId","data-v-bd01e0d5"]]),Ee=s=>(E("data-v-168699b1"),s=s(),D(),s),Xr=Ee(()=>c("div",{class:"curtain"},null,-1)),Zr={class:"nav",id:"VPSidebarNav","aria-labelledby":"sidebar-aria-label",tabindex:"-1"},Qr=Ee(()=>c("span",{class:"visually-hidden",id:"sidebar-aria-label"}," Sidebar Navigation ",-1)),ei=$({__name:"VPSidebar",props:{open:{type:Boolean}},setup(s){const{sidebarGroups:e,hasSidebar:t}=F(),n=s,o=M(null),r=Me(K?document.body:null);return G([n,o],()=>{var d;n.open?(r.value=!0,(d=o.value)==null||d.focus()):r.value=!1},{immediate:!0,flush:"post"}),(d,p)=>i(t)?(a(),l("aside",{key:0,class:T(["VPSidebar",{open:d.open}]),ref_key:"navEl",ref:o,onClick:p[0]||(p[0]=et(()=>{},["stop"]))},[Xr,c("nav",Zr,[Qr,u(d.$slots,"sidebar-nav-before",{},void 0,!0),(a(!0),l(C,null,B(i(e),v=>(a(),l("div",{key:v.text,class:"group"},[_(Jr,{item:v,depth:0},null,8,["item"])]))),128)),u(d.$slots,"sidebar-nav-after",{},void 0,!0)])],2)):f("",!0)}}),ti=m(ei,[["__scopeId","data-v-168699b1"]]),si=$({__name:"VPSkipLink",setup(s){const e=te(),t=M();G(()=>e.path,()=>t.value.focus());function n({target:o}){const r=document.getElementById(decodeURIComponent(o.hash).slice(1));if(r){const d=()=>{r.removeAttribute("tabindex"),r.removeEventListener("blur",d)};r.setAttribute("tabindex","-1"),r.addEventListener("blur",d),r.focus(),window.scrollTo(0,0)}}return(o,r)=>(a(),l(C,null,[c("span",{ref_key:"backToTop",ref:t,tabindex:"-1"},null,512),c("a",{href:"#VPContent",class:"VPSkipLink visually-hidden",onClick:n}," Skip to content ")],64))}}),ni=m(si,[["__scopeId","data-v-c8291ffa"]]),oi=$({__name:"Layout",setup(s){const{isOpen:e,open:t,close:n}=F(),o=te();G(()=>o.path,n),bt(e,n);const{frontmatter:r}=P(),d=tt(),p=b(()=>!!d["home-hero-image"]);return Ce("hero-image-slot-exists",p),(v,g)=>{const w=j("Content");return i(r).layout!==!1?(a(),l("div",{key:0,class:T(["Layout",i(r).pageClass])},[u(v.$slots,"layout-top",{},void 0,!0),_(ni),_(rt,{class:"backdrop",show:i(e),onClick:i(n)},null,8,["show","onClick"]),_(jr,null,{"nav-bar-title-before":h(()=>[u(v.$slots,"nav-bar-title-before",{},void 0,!0)]),"nav-bar-title-after":h(()=>[u(v.$slots,"nav-bar-title-after",{},void 0,!0)]),"nav-bar-content-before":h(()=>[u(v.$slots,"nav-bar-content-before",{},void 0,!0)]),"nav-bar-content-after":h(()=>[u(v.$slots,"nav-bar-content-after",{},void 0,!0)]),"nav-screen-content-before":h(()=>[u(v.$slots,"nav-screen-content-before",{},void 0,!0)]),"nav-screen-content-after":h(()=>[u(v.$slots,"nav-screen-content-after",{},void 0,!0)]),_:3}),_(Zn,{open:i(e),onOpenMenu:i(t)},null,8,["open","onOpenMenu"]),_(ti,{open:i(e)},{"sidebar-nav-before":h(()=>[u(v.$slots,"sidebar-nav-before",{},void 0,!0)]),"sidebar-nav-after":h(()=>[u(v.$slots,"sidebar-nav-after",{},void 0,!0)]),_:3},8,["open"]),_(In,null,{"page-top":h(()=>[u(v.$slots,"page-top",{},void 0,!0)]),"page-bottom":h(()=>[u(v.$slots,"page-bottom",{},void 0,!0)]),"not-found":h(()=>[u(v.$slots,"not-found",{},void 0,!0)]),"home-hero-before":h(()=>[u(v.$slots,"home-hero-before",{},void 0,!0)]),"home-hero-info":h(()=>[u(v.$slots,"home-hero-info",{},void 0,!0)]),"home-hero-image":h(()=>[u(v.$slots,"home-hero-image",{},void 0,!0)]),"home-hero-after":h(()=>[u(v.$slots,"home-hero-after",{},void 0,!0)]),"home-features-before":h(()=>[u(v.$slots,"home-features-before",{},void 0,!0)]),"home-features-after":h(()=>[u(v.$slots,"home-features-after",{},void 0,!0)]),"doc-footer-before":h(()=>[u(v.$slots,"doc-footer-before",{},void 0,!0)]),"doc-before":h(()=>[u(v.$slots,"doc-before",{},void 0,!0)]),"doc-after":h(()=>[u(v.$slots,"doc-after",{},void 0,!0)]),"doc-top":h(()=>[u(v.$slots,"doc-top",{},void 0,!0)]),"doc-bottom":h(()=>[u(v.$slots,"doc-bottom",{},void 0,!0)]),"aside-top":h(()=>[u(v.$slots,"aside-top",{},void 0,!0)]),"aside-bottom":h(()=>[u(v.$slots,"aside-bottom",{},void 0,!0)]),"aside-outline-before":h(()=>[u(v.$slots,"aside-outline-before",{},void 0,!0)]),"aside-outline-after":h(()=>[u(v.$slots,"aside-outline-after",{},void 0,!0)]),"aside-ads-before":h(()=>[u(v.$slots,"aside-ads-before",{},void 0,!0)]),"aside-ads-after":h(()=>[u(v.$slots,"aside-ads-after",{},void 0,!0)]),_:3}),_(Hn),u(v.$slots,"layout-bottom",{},void 0,!0)],2)):(a(),k(w,{key:1}))}}}),ai=m(oi,[["__scopeId","data-v-9d8abc1e"]]),li={Layout:ai,enhanceApp:({app:s})=>{s.component("Badge",nt)}};class ci{constructor(e=10){oe(this,"max");oe(this,"cache");this.max=e,this.cache=new Map}get(e){let t=this.cache.get(e);return t!==void 0&&(this.cache.delete(e),this.cache.set(e,t)),t}set(e,t){this.cache.has(e)?this.cache.delete(e):this.cache.size===this.max&&this.cache.delete(this.first()),this.cache.set(e,t)}first(){return this.cache.keys().next().value}clear(){this.cache.clear()}}export{ci as L,Ta as c,li as t,P as u};
-function __vite__mapDeps(indexes) {
-  if (!__vite__mapDeps.viteFileDeps) {
-    __vite__mapDeps.viteFileDeps = ["assets/chunks/VPLocalSearchBox.5p5Pr0YC.js","assets/chunks/framework.KK7BHIzh.js"]
-  }
-  return indexes.map((i) => __vite__mapDeps.viteFileDeps[i])
-}
