+++ docs/assets/global-flags.md.PwqekeV0.js
@@ -1,7 +1,12 @@
-import{_ as s,o as a,c as i,R as n}from"./chunks/framework.KK7BHIzh.js";const F=JSON.parse('{"title":"Global Flags","description":"","frontmatter":{},"headers":[],"relativePath":"global-flags.md","filePath":"global-flags.md","lastUpdated":1702421137000}'),e={name:"global-flags.md"},t=n(`<h1 id="global-flags" tabindex="-1">Global Flags <a class="header-anchor" href="#global-flags" aria-label="Permalink to &quot;Global Flags&quot;">​</a></h1><p>This page describes the default flags that could be used to alter sana&#39;s behavior.</p><h2 id="profile-p" tabindex="-1"><code>--profile</code>, <code>-p</code> <a class="header-anchor" href="#profile-p" aria-label="Permalink to &quot;\`--profile\`, \`-p\`&quot;">​</a></h2><p>By default, <code>sana</code> utilizes a machine&#39;s default <code>aws-cli</code> profile. To use a <a href="https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html#cli-configure-files-using-profiles" target="_blank" rel="noreferrer">named profile</a>, you can specify the <code>--profile</code> or <code>-p</code> flag.</p><div class="language-sh vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">sh</span><pre class="shiki shiki-themes github-light github-dark vp-code"><code><span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Regular syntax</span></span>
+import{_ as s,o as a,c as i,R as n}from"./chunks/framework.KK7BHIzh.js";const F=JSON.parse('{"title":"Global Flags","description":"","frontmatter":{},"headers":[],"relativePath":"global-flags.md","filePath":"global-flags.md","lastUpdated":1702820404000}'),e={name:"global-flags.md"},l=n(`<h1 id="global-flags" tabindex="-1">Global Flags <a class="header-anchor" href="#global-flags" aria-label="Permalink to &quot;Global Flags&quot;">​</a></h1><p>This page describes the global flags used to alter sana&#39;s behavior.</p><h2 id="profile-p" tabindex="-1"><code>--profile</code>, <code>-p</code> <a class="header-anchor" href="#profile-p" aria-label="Permalink to &quot;\`--profile\`, \`-p\`&quot;">​</a></h2><p>By default, <code>sana</code> utilizes a machine&#39;s default <code>aws-cli</code> profile. To use a <a href="https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html#cli-configure-files-using-profiles" target="_blank" rel="noreferrer">named profile</a>, you can specify the <code>--profile</code> or <code>-p</code> flag.</p><div class="language-sh vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">sh</span><pre class="shiki shiki-themes github-light github-dark vp-code"><code><span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Uses default aws-cli profile</span></span>
+<span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda</span></span>
+<span class="line"></span>
+<span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Regular syntax for using</span></span>
+<span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># named profiles</span></span>
 <span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --profile</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> foo</span></span>
 <span class="line"></span>
-<span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Short-hand syntax</span></span>
+<span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Short-hand syntax for</span></span>
+<span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># named profiles</span></span>
 <span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> -p</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> foo</span></span></code></pre></div><h2 id="region-r" tabindex="-1"><code>--region</code>, <code>-r</code> <a class="header-anchor" href="#region-r" aria-label="Permalink to &quot;\`--region\`, \`-r\`&quot;">​</a></h2><p>By default, <code>sana</code> scans all the regions enabled in an AWS account to provide an account-wide analysis report. If you wish to target regions of interest, you can pass the <code>--region</code> or <code>-r</code> flag to reduce the time required by sana to produce an analysis report.</p><div class="language-sh vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">sh</span><pre class="shiki shiki-themes github-light github-dark vp-code"><code><span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Runs a single-region scan against Singapore region</span></span>
 <span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --region</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> ap-southeast-1</span></span>
 <span class="line"></span>
@@ -23,4 +28,4 @@ import{_ as s,o as a,c as i,R as n}from"./chunks/framework.KK7BHIzh.js";const F=
 <span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda-vpc-distribution</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> -o</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> text</span></span>
 <span class="line"></span>
 <span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># By passing cli-table, you get a well formatted table in the terminal</span></span>
-<span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda-region-distribution</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> -o</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> cli-table</span></span></code></pre></div><p>Valid Inputs:</p><ul><li><code>text</code> (default): prints out results in raw text</li><li><code>chart</code>: prints out a cli-based horizontal chart for distribution commands</li><li><code>cli-table</code>: prints out a cli-based table for distribution commands.</li></ul><h2 id="no-progress-bar" tabindex="-1"><code>--no-progress-bar</code> <a class="header-anchor" href="#no-progress-bar" aria-label="Permalink to &quot;\`--no-progress-bar\`&quot;">​</a></h2><p>You can use the <code>--no-progress-bar</code> flag to disable the progress bar.</p><div class="language-sh vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">sh</span><pre class="shiki shiki-themes github-light github-dark vp-code"><code><span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda-count</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --no-progress-bar</span></span></code></pre></div><blockquote><p>TIP: This flag is often used in automation servers.</p></blockquote><h2 id="silent-mode" tabindex="-1"><code>--silent-mode</code> <a class="header-anchor" href="#silent-mode" aria-label="Permalink to &quot;\`--silent-mode\`&quot;">​</a></h2><p>You can use the <code>--silent-mode</code> mode flag to prevent non-critical logs from getting displayed</p><div class="language-sh vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">sh</span><pre class="shiki shiki-themes github-light github-dark vp-code"><code><span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda-avg-package-size</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --silent-mode</span></span></code></pre></div>`,20),l=[t];function h(p,o,r,d,k,c){return a(),i("div",null,l)}const u=s(e,[["render",h]]);export{F as __pageData,u as default};
+<span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda-region-distribution</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> -o</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> cli-table</span></span></code></pre></div><p>Valid Inputs:</p><ul><li><code>text</code> (default): prints out results in raw text</li><li><code>chart</code>: prints out a cli-based horizontal chart for distribution commands</li><li><code>cli-table</code>: prints out a cli-based table for distribution commands.</li></ul><h2 id="no-progress-bar" tabindex="-1"><code>--no-progress-bar</code> <a class="header-anchor" href="#no-progress-bar" aria-label="Permalink to &quot;\`--no-progress-bar\`&quot;">​</a></h2><p>You can use the <code>--no-progress-bar</code> flag to disable the progress bar.</p><div class="language-sh vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">sh</span><pre class="shiki shiki-themes github-light github-dark vp-code"><code><span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda-count</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --no-progress-bar</span></span></code></pre></div><blockquote><p>TIP: This flag is often used in automation servers.</p></blockquote><h2 id="silent-mode" tabindex="-1"><code>--silent-mode</code> <a class="header-anchor" href="#silent-mode" aria-label="Permalink to &quot;\`--silent-mode\`&quot;">​</a></h2><p>You can use the <code>--silent-mode</code> mode flag to prevent non-critical logs from getting displayed</p><div class="language-sh vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">sh</span><pre class="shiki shiki-themes github-light github-dark vp-code"><code><span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">$</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> sana</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> lambda-avg-package-size</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --silent-mode</span></span></code></pre></div>`,20),t=[l];function p(h,o,r,d,k,c){return a(),i("div",null,t)}const u=s(e,[["render",p]]);export{F as __pageData,u as default};
