--- docs/assets/index.md._HVROeOM.lean.js
@@ -1,1 +1,0 @@
-import{_ as e,o as t,c as n}from"./chunks/framework.KK7BHIzh.js";const m=JSON.parse(`{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"sana","text":"Analyze your AWS account's serverless resources","tagline":"Designed by AWS serverless engineers, for AWS serverless engineers","actions":[{"theme":"brand","text":"Get Started","link":"/get-started"},{"theme":"alt","text":"View on Github","link":"https://github.com/allanchua101/sana"}]},"features":[{"title":"Account-wide Scanning","details":"Rapidly scan an entire account's serverless resources without leaving your VSCode"},{"title":"Distribution Analysis","details":"Uncover misconfigurations, outliers and anomalies using distribution analysis."},{"title":"Modern Web-based Documentation","details":"Easy to read web-based documentation built on top of modern tooling."}]},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":1702775615000}`),a={name:"index.md"};function s(i,o,r,d,l,c){return t(),n("div")}const p=e(a,[["render",s]]);export{m as __pageData,p as default};
