+++ docs/assets/index.md.oCVJZTGS.js
@@ -1,0 +1,1 @@
+import{_ as e,o as t,c as a}from"./chunks/framework.KK7BHIzh.js";const m=JSON.parse(`{"title":"sana","description":"","frontmatter":{"layout":"home","title":"sana","desc":"Analyze your AWS account's serverless resources","hero":{"name":"sana","text":"Analyze your AWS account's serverless resources","tagline":"Made by frugal engineers, for AWS serverless engineers.","actions":[{"theme":"brand","text":"Get Started","link":"/get-started"},{"theme":"alt","text":"View on Github","link":"https://github.com/allanchua101/sana"}]},"features":[{"icon":"🌐","title":"Account-wide Scanning","details":"Rapidly scan an entire account's serverless resources without leaving your VSCode"},{"icon":"📊","title":"Distribution Analysis","details":"Uncover misconfigurations, outliers and anomalies using distribution analysis."},{"icon":"⚙️","title":"Modern Web-based Documentation","details":"Easy to read web-based documentation built on top of modern tooling."}]},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":1702775615000}`),n={name:"index.md"};function s(o,i,r,c,l,d){return t(),a("div")}const p=e(n,[["render",s]]);export{m as __pageData,p as default};
